### 2025-03-06 #7872 @bsun1220
> [@bsun1220](https://github.com/bsun1220) if you have any suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @). You can search [the list of people](https://reviewers.joss.theoj.org/lookup) that have already agreed to review and may be suitable for this submission.

Using the reviewer suggestions from [matt-graham](https://github.com/matt-graham) who has a different paper on MCMC algorithms, @vissarion if you need help finding reviewers, I saw that [larryshamalama](https://github.com/larryshamalama), [vankesteren](https://github.com/vankesteren), [robmoss](https://github.com/robmoss), [njtierney](https://github.com/njtierney) and [martinmodrak](https://github.com/martinmodrak) all had relevant profiles.
