+++ .github/workflows/ciwheels.yml
@@ -3,9 +3,7 @@ name: Build CI wheels
 on:
   push:
     branches:
-      - main
-    tags:
-      - 'v*'  # Triggers on tag pushes that match the pattern (e.g., v1.0.0, v2.0.1, etc.)
+      - log-concave
   pull_request:
     branches:
       - main
@@ -106,57 +104,3 @@ jobs:
         with:
           name: cibw-wheels-${{ matrix.os }}-cp${{ matrix.python }}-${{ matrix.platform_id }}
           path: ./wheelhouse/*.whl
-
-  publish-to-pypi:
-    name: >- 
-      Publish Python 🐍 distribution 📦 to PyPI
-    if: startsWith(github.ref, 'refs/tags/')  # only publish on tag pushes
-    needs:
-    - build_wheels
-    runs-on: ubuntu-latest
-
-    environment:
-      name: pypi
-      url: https://pypi.org/p/polytopewalk
-
-    permissions:
-      id-token: write  # IMPORTANT: mandatory for trusted publishing
-
-    steps:
-    - name: Download all the wheels
-      uses: actions/download-artifact@v4
-      with:
-        pattern: cibw-*
-        path: dist
-        merge-multiple: true
-
-    - name: Publish distribution 📦 to PyPI
-      uses: pypa/gh-action-pypi-publish@release/v1
-
-  publish-to-testpypi:
-    name: >- 
-      Publish Python 🐍 distribution 📦 to TestPyPI
-    if: startsWith(github.ref, 'refs/tags/')  # only publish on tag pushes
-    needs:
-    - build_wheels
-    runs-on: ubuntu-latest
-
-    environment:
-      name: testpypi
-      url: https://test.pypi.org/p/polytopewalk
-
-    permissions:
-      id-token: write  # IMPORTANT: mandatory for trusted publishing
-
-    steps:
-    - name: Download all the wheels
-      uses: actions/download-artifact@v4
-      with:
-        pattern: cibw-*
-        path: dist
-        merge-multiple: true
-
-    - name: Publish distribution 📦 to TestPyPI
-      uses: pypa/gh-action-pypi-publish@release/v1
-      with:
-        repository-url: https://test.pypi.org/legacy/
