### 2025-03-27 #7872 @bsun1220
Hey @vissarion,

Additionally, I think [matt-graham](https://github.com/matt-graham), who is currently submitting a package on MCMC sampling in R in JOSS, could be a relevant potential reviewer as well. Thanks!
