### 2025-03-28 #7872 @vissarion
Hi, @matt-graham do you have some time to review this submission for JOSS? Your background is highly relevant.
