### 2025-03-31 #7957 @editorialbot
[REVIEW]: PolytopeWalk: Sparse MCMC Sampling over Polytopes
**Submitting author:** @bsun1220 (<a href="http://orcid.org/0009-0005-9635-4800">Benny Sun</a>)
**Repository:** https://github.com/ethz-randomwalk/polytopewalk
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.17
