+++ .github/workflows/ciwheels.yml
@@ -75,7 +75,7 @@ jobs:
             platform_id: macosx_arm64
             deployment_target: "14"
           - os: macos-latest
-            python: 312
+            python: 313
             platform_id: macosx_arm64
             deployment_target: "14"
        
