+++ .github/workflows/ciwheels.yml
@@ -83,15 +83,15 @@ jobs:
       - name: Checkout code
         uses: actions/checkout@v4
       
-      - name: Install packages (Windows)
-        if: runner.os == 'Windows'
-        shell: bash
-        run: |
-          curl -owinglpk-4.65.zip -L --insecure https://jaist.dl.sourceforge.net/project/winglpk/winglpk/GLPK-4.65/winglpk-4.65.zip
-          7z x winglpk-4.65.zip
-          cp glpk-4.65/w64/glpk_4_65.lib glpk-4.65/w64/glpk.lib
-          echo GLPK_LIB_DIR=${GITHUB_WORKSPACE}\\glpk-4.65\\w64 >> $GITHUB_ENV
-          echo GLPK_INCLUDE_DIR=${GITHUB_WORKSPACE}\\glpk-4.65\\src >> $GITHUB_ENV
+      # - name: Install packages (Windows)
+      #   if: runner.os == 'Windows'
+      #   shell: bash
+      #   run: |
+      #     curl -owinglpk-4.65.zip -L --insecure https://jaist.dl.sourceforge.net/project/winglpk/winglpk/GLPK-4.65/winglpk-4.65.zip
+      #     7z x winglpk-4.65.zip
+      #     cp glpk-4.65/w64/glpk_4_65.lib glpk-4.65/w64/glpk.lib
+      #     echo GLPK_LIB_DIR=${GITHUB_WORKSPACE}\\glpk-4.65\\w64 >> $GITHUB_ENV
+      #     echo GLPK_INCLUDE_DIR=${GITHUB_WORKSPACE}\\glpk-4.65\\src >> $GITHUB_ENV
           
       - name: Build wheels
         uses: pypa/cibuildwheel@v2.20.0
@@ -107,7 +107,7 @@ jobs:
             yum install -y epel-release eigen3-devel glpk-devel
           CIBW_BEFORE_ALL_WINDOWS: >
             echo "OK starts Windows build" &&
-            choco install eigen -y
+            choco install eigen glpk -y
         with:
           package-dir: .
           output-dir: wheelhouse
