### 2025-04-22 384d5f7
added MANIFEST.in to include dll
