+++ requirements.txt
@@ -1,4 +1,3 @@
 numpy
 pandas
 scipy
-polytopewalk
