+++ .github/workflows/test.yml
@@ -25,6 +25,7 @@ jobs:
         run: |
           python -m pip install --upgrade pip
           pip install -r requirements.txt
+          pip install .
 
       - name: Run tests
         run: python -m unittest discover -s tests/python -p "*.py"
