### 2025-04-22 be896e6
added glpk & eigen
