### 2025-04-22 9b0eba1
removed MANIFEST.in, not useful for dll
