### 2025-04-22 c1262d3
add-path to find glpk to repair
