+++ .github/workflows/ciwheels.yml
@@ -110,7 +110,7 @@ jobs:
             echo "OK starts Windows build" &&
             choco install eigen -y
           CIBW_BEFORE_BUILD_WINDOWS: "pip install delvewheel"
-          CIBW_REPAIR_WHEEL_COMMAND_WINDOWS: "delvewheel repair -w {dest_dir} {wheel}"
+          CIBW_REPAIR_WHEEL_COMMAND_WINDOWS: "delvewheel repair --add-path \"glpk-4.65/w64/\" -w {dest_dir} {wheel}"
         with:
           package-dir: .
           output-dir: wheelhouse
