### 2025-04-22 bedd017
added scikit-build.resources
