+++ .github/workflows/test.yml
@@ -34,8 +34,6 @@ jobs:
           python -m pip install --upgrade pip
           pip install -r requirements.txt
           pip install .
-          pip install delvewheel
-          delvewheel repair --add-path \"glpk-4.65/w64/\" -w {dest_dir} {wheel}
 
       - name: Run tests
         run: python -m unittest discover -s tests/python -p "*.py"
