### 2025-04-22 c7e8199
install DLL via Cmake
