+++ .github/workflows/ciwheels.yml
@@ -91,6 +91,7 @@ jobs:
           7z x winglpk-4.65.zip
           cp glpk-4.65/w64/glpk_4_65.lib glpk-4.65/w64/glpk.lib
           cp glpk-4.65/w64/glpk_4_65.dll glpk-4.65/w64/glpk.dll
+          cp glpk-4.65/w64/glpk_4_65.dll lib/glpk_4_65.dll
           echo GLPK_LIB_DIR=${GITHUB_WORKSPACE}\\glpk-4.65\\w64 >> $GITHUB_ENV
           echo GLPK_INCLUDE_DIR=${GITHUB_WORKSPACE}\\glpk-4.65\\src >> $GITHUB_ENV
           
