+++ pyproject.toml
@@ -53,13 +53,6 @@ build.verbose = true
 # "Release", "RelWithDebInfo", "MinSizeRel", "", etc.
 cmake.build-type = "Release"
 
-include-package-data = true
-
-[tool.scikit-build.resources]
-platform = "windows"
-path = "**/*.dll"
-src = "glpk-4.65/w64"
-
 
 [tool.scikit-build-core.env]
 # Allow overriding CMake options via environment variables
