+++ CMakeLists.txt
@@ -42,11 +42,11 @@ target_link_libraries(${PROJECT_NAME} PUBLIC utils dense sparse pybind11::module
 target_include_directories(${PROJECT_NAME} PRIVATE ${Eigen3_INCLUDE_DIRS} ${GLPK_INCLUDE_DIR})
 install(TARGETS ${PROJECT_NAME} LIBRARY DESTINATION ${PROJECT_NAME})
 
-# Windows-specific DLL installation
-if(WIN32)
-    file(GLOB DLL_FILES "${CMAKE_CURRENT_SOURCE_DIR}/lib/*.dll")
-    install(FILES ${DLL_FILES} DESTINATION ${PROJECT_NAME})
-endif()
+# # Windows-specific DLL installation
+# if(WIN32)
+#     file(GLOB DLL_FILES "${CMAKE_CURRENT_SOURCE_DIR}/lib/*.dll")
+#     install(FILES ${DLL_FILES} DESTINATION ${PROJECT_NAME})
+# endif()
 
 # if (BUILD_TESTS)
 add_subdirectory("tests")
