### 2025-04-22 728bfdd
fixed cmake loc, prepare for 1.0.23
