### 2025-04-22 08b444c
added mac-os-13 x86_64 to the list of supported platforms, prepare for 1.0.24
