+++ .github/workflows/ciwheels.yml
@@ -57,6 +57,27 @@ jobs:
             manylinux_image: manylinux2014
   
           # MacOS macos-12 x86_64 is deprecated
+          # Macos macos-13 x86_64
+          - os: macos-13
+            python: 39
+            platform_id: macosx_x86_64
+            deployment_target: "13"
+          - os: macos-13
+            python: 310
+            platform_id: macosx_x86_64
+            deployment_target: "13"
+          - os: macos-13
+            python: 311
+            platform_id: macosx_x86_64
+            deployment_target: "13"
+          - os: macos-13
+            python: 312
+            platform_id: macosx_x86_64
+            deployment_target: "13"
+          - os: macos-13
+            python: 313
+            platform_id: macosx_x86_64
+            deployment_target: "13"
           # MacOS macos-14 arm64
           - os: macos-latest
             python: 39
