### 2025-05-09 ethz-randomwalk/polytopewalk#8 @matt-graham
Python dependencies not specified in project metadata
Raising as part of openjournals/joss-reviews#7957

The Python dependencies of the package are not currently specified in the `pyproject.toml` file. Specifying them there under [`project.dependencies`](https://packaging.python.org/en/latest/guides/writing-pyproject-toml/#dependencies-and-requirements) along with any version constraints would allow users to install the package with a single `pip install`  command without having to separately install the dependencies, and also mean if other projects specify  `polytopewalk` as a dependency, that `polytopewalk`'s dependencies are automatically resolved and installed and any conflict with requirements of other packages flagged.

Similarly it would be helpful to specify the minimum Python version supported in `pyproject.toml` [by specifying the `project.requires-python` field](https://packaging.python.org/en/latest/guides/writing-pyproject-toml/#python-requires).

Specifically the dependencies currently specified in the `README` file could be defined in the `pyproject.toml` file by adding the following text

```toml
[project]
dependencies = [
  "numpy>=1.20",
  "scipy>=1.6.0",
]
requires-python = ">= 3.9"
```
