### 2025-05-09 ethz-randomwalk/polytopewalk#9 @matt-graham
Equations in README rendered as images which impedes readability
Raising as part of openjournals/joss-reviews#7957

Two of the equations in the README are rendered as images using the CodeCogs Equation Rendering webservice:

https://github.com/ethz-randomwalk/polytopewalk/blob/55fd336c348b5f3724079534804a4d9eca2301b0/README.md?plain=1#L34
https://github.com/ethz-randomwalk/polytopewalk/blob/55fd336c348b5f3724079534804a4d9eca2301b0/README.md?plain=1#L40

From an accessibility point of view using [GitHub built-in maths support in Markdown](https://docs.github.com/en/get-started/writing-on-github/working-with-advanced-formatting/writing-mathematical-expressions) would be better, as the underlying MathJax engine used to render equations provides screenreader support. The current images also are not very clear to read when viewing the README in a dark-mode theme, for example see image below:

![Screenshot of equation in README in dark-mode](https://github.com/user-attachments/assets/f1b40cd8-8be9-41ca-84b5-1de1a74727b9)
