### 2025-05-09 ethz-randomwalk/polytopewalk#10 @matt-graham
Unable to install Python package from source
Raising as part of openjournals/joss-reviews#7957

When attempting to install the package from source on a Ubuntu (24.04) system I get build errors

<summary>Error output</summary>

```
Processing /polytopewalk
  Installing build dependencies ... done
  Installing backend dependencies ... done
  Preparing metadata (pyproject.toml) ... done
Building wheels for collected packages: polytopewalk
  Building wheel for polytopewalk (pyproject.toml) ... error
  error: subprocess-exited-with-error

  × Building wheel for polytopewalk (pyproject.toml) did not run successfully.
  │ exit code: 1
  ╰─> [132 lines of output]
      *** scikit-build-core 0.11.2 using CMake 4.0.2 (wheel)
      *** Configuring CMake...
      loading initial cache file /tmp/tmplbg1x2ny/build/CMakeInit.txt
      -- The C compiler identification is GNU 13.3.0
      -- The CXX compiler identification is GNU 13.3.0
      -- Detecting C compiler ABI info
      -- Detecting C compiler ABI info - done
      -- Check for working C compiler: /usr/bin/gcc - skipped
      -- Detecting C compile features
      -- Detecting C compile features - done
      -- Detecting CXX compiler ABI info
      -- Check for working CXX compiler: /usr/bin/g++ - skipped
      -- Detecting CXX compile features
      -- CMAKE_MODULE_PATH: /polytopewalk/cmake/Modules/
      -- Found Python: /polytopewalk/.venv/bin/python3 (found version "3.12.3") found components: Interpreter Development.Module
      -- Performing Test HAS_FLTO
      -- Performing Test HAS_FLTO - Success
      -- Found pybind11: /tmp/pip-build-env-8r5owqgk/overlay/lib/python3.12/site-packages/pybind11/include (found version "2.13.6")
      -- Found GLPK: /usr/lib/x86_64-linux-gnu/libglpk.so
      CMake Deprecation Warning at /tmp/tmplbg1x2ny/build/_deps/catch2-src/CMakeLists.txt:1 (cmake_minimum_required):
        Compatibility with CMake < 3.10 will be removed from a future version of
        CMake.

        Update the VERSION argument <min> value.  Or, use the <min>...<max> syntax
        to tell CMake that the project requires at least <min> but has been updated
        to work with policies introduced by <max> or earlier.

      -- Configuring done (9.8s)
      -- Generating done (0.1s)
      -- Build files have been written to: /tmp/tmplbg1x2ny/build
      *** Building project with Ninja...
      Change Dir: '/tmp/tmplbg1x2ny/build'

      Run Build Command(s): ninja -v
      [1/115] /usr/bin/g++  -I/polytopewalk/src/. -I/polytopewalk/src/Eigen3::Eigen -isystem /usr/include/eigen3 -O2 -fPIC -O3 -DNDEBUG -std=gnu++11 -MD -MT src/CMakeFiles/dense.dir/dense/DikinWalk.cpp.o -MF src/CMakeFiles/dense.dir/dense/DikinWalk.cpp.o.d -o src/CMakeFiles/dense.dir/dense/DikinWalk.cpp.o -c /polytopewalk/src/dense/DikinWalk.cpp
      [2/115] /usr/bin/g++  -I/polytopewalk/src/. -I/polytopewalk/src/Eigen3::Eigen -isystem /usr/include/eigen3 -O2 -fPIC -O3 -DNDEBUG -std=gnu++11 -MD -MT src/CMakeFiles/dense.dir/dense/BallWalk.cpp.o -MF src/CMakeFiles/dense.dir/dense/BallWalk.cpp.o.d -o src/CMakeFiles/dense.dir/dense/BallWalk.cpp.o -c /polytopewalk/src/dense/BallWalk.cpp
      [3/115] /usr/bin/g++  -I/polytopewalk/src/. -I/polytopewalk/src/Eigen3::Eigen -isystem /usr/include/eigen3 -O2 -fPIC -O3 -DNDEBUG -std=gnu++11 -MD -MT src/CMakeFiles/dense.dir/dense/RandomWalk.cpp.o -MF src/CMakeFiles/dense.dir/dense/RandomWalk.cpp.o.d -o src/CMakeFiles/dense.dir/dense/RandomWalk.cpp.o -c /polytopewalk/src/dense/RandomWalk.cpp
      [4/115] /usr/bin/g++  -I/polytopewalk/src/. -I/polytopewalk/src/Eigen3::Eigen -isystem /usr/include/eigen3 -O2 -fPIC -O3 -DNDEBUG -std=gnu++11 -MD -MT src/CMakeFiles/dense.dir/dense/HitRun.cpp.o -MF src/CMakeFiles/dense.dir/dense/HitRun.cpp.o.d -o src/CMakeFiles/dense.dir/dense/HitRun.cpp.o -c /polytopewalk/src/dense/HitRun.cpp
      [5/115] /usr/bin/g++  -I/polytopewalk/src/. -I/polytopewalk/src/Eigen3::Eigen -isystem /usr/include/eigen3 -O2 -fPIC -O3 -DNDEBUG -std=gnu++11 -MD -MT src/CMakeFiles/sparse.dir/sparse/SparseDikinWalk.cpp.o -MF src/CMakeFiles/sparse.dir/sparse/SparseDikinWalk.cpp.o.d -o src/CMakeFiles/sparse.dir/sparse/SparseDikinWalk.cpp.o -c /polytopewalk/src/sparse/SparseDikinWalk.cpp
      [6/115] /usr/bin/g++  -I/polytopewalk/src/. -I/polytopewalk/src/Eigen3::Eigen -isystem /usr/include/eigen3 -O2 -fPIC -O3 -DNDEBUG -std=gnu++11 -MD -MT src/CMakeFiles/sparse.dir/sparse/SparseDikinLSWalk.cpp.o -MF src/CMakeFiles/sparse.dir/sparse/SparseDikinLSWalk.cpp.o.d -o src/CMakeFiles/sparse.dir/sparse/SparseDikinLSWalk.cpp.o -c /polytopewalk/src/sparse/SparseDikinLSWalk.cpp
      [7/115] /usr/bin/g++  -I/polytopewalk/src/. -I/polytopewalk/src/Eigen3::Eigen -isystem /usr/include/eigen3 -O2 -fPIC -O3 -DNDEBUG -std=gnu++11 -MD -MT src/CMakeFiles/sparse.dir/sparse/LeverageScore.cpp.o -MF src/CMakeFiles/sparse.dir/sparse/LeverageScore.cpp.o.d -o src/CMakeFiles/sparse.dir/sparse/LeverageScore.cpp.o -c /polytopewalk/src/sparse/LeverageScore.cpp
      [8/115] /usr/bin/g++  -I/polytopewalk/src/. -I/polytopewalk/src/Eigen3::Eigen -isystem /usr/include/eigen3 -O2 -fPIC -O3 -DNDEBUG -std=gnu++11 -MD -MT src/CMakeFiles/sparse.dir/sparse/SparseRandomWalk.cpp.o -MF src/CMakeFiles/sparse.dir/sparse/SparseRandomWalk.cpp.o.d -o src/CMakeFiles/sparse.dir/sparse/SparseRandomWalk.cpp.o -c /polytopewalk/src/sparse/SparseRandomWalk.cpp
      [9/115] /usr/bin/g++  -I/polytopewalk/src/. -I/polytopewalk/src/Eigen3::Eigen -isystem /usr/include/eigen3 -O2 -fPIC -O3 -DNDEBUG -std=gnu++11 -MD -MT src/CMakeFiles/dense.dir/dense/VaidyaWalk.cpp.o -MF src/CMakeFiles/dense.dir/dense/VaidyaWalk.cpp.o.d -o src/CMakeFiles/dense.dir/dense/VaidyaWalk.cpp.o -c /polytopewalk/src/dense/VaidyaWalk.cpp
      [10/115] /usr/bin/g++  -I/polytopewalk/src/. -I/polytopewalk/src/Eigen3::Eigen -isystem /usr/include/eigen3 -O2 -fPIC -O3 -DNDEBUG -std=gnu++11 -MD -MT src/CMakeFiles/sparse.dir/sparse/SparseJohnWalk.cpp.o -MF src/CMakeFiles/sparse.dir/sparse/SparseJohnWalk.cpp.o.d -o src/CMakeFiles/sparse.dir/sparse/SparseJohnWalk.cpp.o -c /polytopewalk/src/sparse/SparseJohnWalk.cpp
      [11/115] /usr/bin/g++  -I/polytopewalk/src/. -I/polytopewalk/src/Eigen3::Eigen -isystem /usr/include/eigen3 -O2 -fPIC -O3 -DNDEBUG -std=gnu++11 -MD -MT src/CMakeFiles/dense.dir/dense/BarrierWalk.cpp.o -MF src/CMakeFiles/dense.dir/dense/BarrierWalk.cpp.o.d -o src/CMakeFiles/dense.dir/dense/BarrierWalk.cpp.o -c /polytopewalk/src/dense/BarrierWalk.cpp
      [12/115] /usr/bin/g++  -I/polytopewalk/src/. -I/polytopewalk/src/Eigen3::Eigen -isystem /usr/include/eigen3 -O2 -fPIC -O3 -DNDEBUG -std=gnu++11 -MD -MT src/CMakeFiles/dense.dir/dense/JohnWalk.cpp.o -MF src/CMakeFiles/dense.dir/dense/JohnWalk.cpp.o.d -o src/CMakeFiles/dense.dir/dense/JohnWalk.cpp.o -c /polytopewalk/src/dense/JohnWalk.cpp
      [13/115] /usr/bin/g++  -I/polytopewalk/src/. -I/polytopewalk/src/Eigen3::Eigen -isystem /usr/include/eigen3 -O2 -fPIC -O3 -DNDEBUG -std=gnu++11 -MD -MT src/CMakeFiles/dense.dir/dense/DikinLSWalk.cpp.o -MF src/CMakeFiles/dense.dir/dense/DikinLSWalk.cpp.o.d -o src/CMakeFiles/dense.dir/dense/DikinLSWalk.cpp.o -c /polytopewalk/src/dense/DikinLSWalk.cpp
      [14/115] : && /tmp/pip-build-env-8r5owqgk/normal/lib/python3.12/site-packages/cmake/data/bin/cmake -E rm -f src/libdense.a && /usr/bin/ar qc src/libdense.a  src/CMakeFiles/dense.dir/dense/BallWalk.cpp.o src/CMakeFiles/dense.dir/dense/BarrierWalk.cpp.o src/CMakeFiles/dense.dir/dense/DikinLSWalk.cpp.o src/CMakeFiles/dense.dir/dense/DikinWalk.cpp.o src/CMakeFiles/dense.dir/dense/HitRun.cpp.o src/CMakeFiles/dense.dir/dense/JohnWalk.cpp.o src/CMakeFiles/dense.dir/dense/RandomWalk.cpp.o src/CMakeFiles/dense.dir/dense/VaidyaWalk.cpp.o && /usr/bin/ranlib src/libdense.a && :
      [15/115] /usr/bin/g++  -I/polytopewalk/src/. -I/polytopewalk/src/Eigen3::Eigen -isystem /usr/include/eigen3 -O2 -fPIC -O3 -DNDEBUG -std=gnu++11 -MD -MT src/CMakeFiles/sparse.dir/sparse/SparseVaidyaWalk.cpp.o -MF src/CMakeFiles/sparse.dir/sparse/SparseVaidyaWalk.cpp.o.d -o src/CMakeFiles/sparse.dir/sparse/SparseVaidyaWalk.cpp.o -c /polytopewalk/src/sparse/SparseVaidyaWalk.cpp
      [16/115] /usr/bin/g++  -I/polytopewalk/src/. -I/polytopewalk/src/Eigen3::Eigen -isystem /usr/include/eigen3 -O2 -fPIC -O3 -DNDEBUG -std=gnu++11 -MD -MT src/CMakeFiles/utils.dir/utils/DenseCenter.cpp.o -MF src/CMakeFiles/utils.dir/utils/DenseCenter.cpp.o.d -o src/CMakeFiles/utils.dir/utils/DenseCenter.cpp.o -c /polytopewalk/src/utils/DenseCenter.cpp
      [17/115] /usr/bin/g++  -I/polytopewalk/src/. -I/polytopewalk/src/Eigen3::Eigen -isystem /usr/include/eigen3 -O2 -fPIC -O3 -DNDEBUG -std=gnu++11 -MD -MT src/CMakeFiles/utils.dir/utils/SparseLP.cpp.o -MF src/CMakeFiles/utils.dir/utils/SparseLP.cpp.o.d -o src/CMakeFiles/utils.dir/utils/SparseLP.cpp.o -c /polytopewalk/src/utils/SparseLP.cpp
      [18/115] /usr/bin/g++  -I/polytopewalk/src/. -I/polytopewalk/src/Eigen3::Eigen -isystem /usr/include/eigen3 -O2 -fPIC -O3 -DNDEBUG -std=gnu++11 -MD -MT src/CMakeFiles/utils.dir/utils/SparseCenter.cpp.o -MF src/CMakeFiles/utils.dir/utils/SparseCenter.cpp.o.d -o src/CMakeFiles/utils.dir/utils/SparseCenter.cpp.o -c /polytopewalk/src/utils/SparseCenter.cpp
      [19/115] /usr/bin/g++ -Dpolytopewalk_EXPORTS -I/polytopewalk/src/. -I/polytopewalk/src/Eigen3::Eigen -isystem /usr/include/python3.12 -isystem /tmp/pip-build-env-8r5owqgk/overlay/lib/python3.12/site-packages/pybind11/include -isystem /usr/include/eigen3 -O2 -fPIC -O3 -DNDEBUG -std=gnu++11 -fPIC -fvisibility=hidden -flto -fno-fat-lto-objects -MD -MT CMakeFiles/polytopewalk.dir/pybind11/PybindExt.cpp.o -MF CMakeFiles/polytopewalk.dir/pybind11/PybindExt.cpp.o.d -o CMakeFiles/polytopewalk.dir/pybind11/PybindExt.cpp.o -c /polytopewalk/pybind11/PybindExt.cpp
      [20/115] /usr/bin/g++  -I/tmp/tmplbg1x2ny/build/_deps/catch2-src/src/catch2/.. -I/tmp/tmplbg1x2ny/build/generated-includes -O2 -fPIC -O3 -DNDEBUG -std=gnu++14 -ffile-prefix-map=/tmp/tmplbg1x2ny/build/_deps/catch2-src=. -MD -MT _deps/catch2-build/src/CMakeFiles/Catch2.dir/catch2/reporters/catch_reporter_automake.cpp.o -MF _deps/catch2-build/src/CMakeFiles/Catch2.dir/catch2/reporters/catch_reporter_automake.cpp.o.d -o _deps/catch2-build/src/CMakeFiles/Catch2.dir/catch2/reporters/catch_reporter_automake.cpp.o -c /tmp/tmplbg1x2ny/build/_deps/catch2-src/src/catch2/reporters/catch_reporter_automake.cpp
      [21/115] /usr/bin/g++  -I/tmp/tmplbg1x2ny/build/_deps/catch2-src/src/catch2/.. -I/tmp/tmplbg1x2ny/build/generated-includes -O2 -fPIC -O3 -DNDEBUG -std=gnu++14 -ffile-prefix-map=/tmp/tmplbg1x2ny/build/_deps/catch2-src=. -MD -MT _deps/catch2-build/src/CMakeFiles/Catch2.dir/catch2/reporters/catch_reporter_combined_tu.cpp.o -MF _deps/catch2-build/src/CMakeFiles/Catch2.dir/catch2/reporters/catch_reporter_combined_tu.cpp.o.d -o _deps/catch2-build/src/CMakeFiles/Catch2.dir/catch2/reporters/catch_reporter_combined_tu.cpp.o -c /tmp/tmplbg1x2ny/build/_deps/catch2-src/src/catch2/reporters/catch_reporter_combined_tu.cpp
      [22/115] /usr/bin/g++  -I/polytopewalk/src/. -I/polytopewalk/src/Eigen3::Eigen -isystem /usr/include/eigen3 -O2 -fPIC -O3 -DNDEBUG -std=gnu++11 -MD -MT src/CMakeFiles/sparse.dir/sparse/SparseBallWalk.cpp.o -MF src/CMakeFiles/sparse.dir/sparse/SparseBallWalk.cpp.o.d -o src/CMakeFiles/sparse.dir/sparse/SparseBallWalk.cpp.o -c /polytopewalk/src/sparse/SparseBallWalk.cpp
      [23/115] /usr/bin/g++  -I/tmp/tmplbg1x2ny/build/_deps/catch2-src/src/catch2/.. -I/tmp/tmplbg1x2ny/build/generated-includes -O2 -fPIC -O3 -DNDEBUG -std=gnu++14 -ffile-prefix-map=/tmp/tmplbg1x2ny/build/_deps/catch2-src=. -MD -MT _deps/catch2-build/src/CMakeFiles/Catch2.dir/catch2/reporters/catch_reporter_common_base.cpp.o -MF _deps/catch2-build/src/CMakeFiles/Catch2.dir/catch2/reporters/catch_reporter_common_base.cpp.o.d -o _deps/catch2-build/src/CMakeFiles/Catch2.dir/catch2/reporters/catch_reporter_common_base.cpp.o -c /tmp/tmplbg1x2ny/build/_deps/catch2-src/src/catch2/reporters/catch_reporter_common_base.cpp
      [24/115] /usr/bin/g++  -I/polytopewalk/src/. -I/polytopewalk/src/Eigen3::Eigen -I/tmp/tmplbg1x2ny/build/_deps/catch2-src/src/catch2/.. -I/tmp/tmplbg1x2ny/build/generated-includes -isystem /usr/include/eigen3 -O2 -fPIC -O3 -DNDEBUG -std=gnu++14 -MD -MT tests/CMakeFiles/test_dense_walk.dir/cpp/test_dense_walk.cpp.o -MF tests/CMakeFiles/test_dense_walk.dir/cpp/test_dense_walk.cpp.o.d -o tests/CMakeFiles/test_dense_walk.dir/cpp/test_dense_walk.cpp.o -c /polytopewalk/tests/cpp/test_dense_walk.cpp
      [25/115] /usr/bin/g++  -I/tmp/tmplbg1x2ny/build/_deps/catch2-src/src/catch2/.. -I/tmp/tmplbg1x2ny/build/generated-includes -O2 -fPIC -O3 -DNDEBUG -std=gnu++14 -ffile-prefix-map=/tmp/tmplbg1x2ny/build/_deps/catch2-src=. -MD -MT _deps/catch2-build/src/CMakeFiles/Catch2.dir/catch2/reporters/catch_reporter_compact.cpp.o -MF _deps/catch2-build/src/CMakeFiles/Catch2.dir/catch2/reporters/catch_reporter_compact.cpp.o.d -o _deps/catch2-build/src/CMakeFiles/Catch2.dir/catch2/reporters/catch_reporter_compact.cpp.o -c /tmp/tmplbg1x2ny/build/_deps/catch2-src/src/catch2/reporters/catch_reporter_compact.cpp
      [26/115] /usr/bin/g++  -I/polytopewalk/src/. -I/polytopewalk/src/Eigen3::Eigen -I/tmp/tmplbg1x2ny/build/_deps/catch2-src/src/catch2/.. -I/tmp/tmplbg1x2ny/build/generated-includes -isystem /usr/include/eigen3 -O2 -fPIC -O3 -DNDEBUG -std=gnu++14 -MD -MT tests/CMakeFiles/test_sparse_walk.dir/cpp/test_sparse_walk.cpp.o -MF tests/CMakeFiles/test_sparse_walk.dir/cpp/test_sparse_walk.cpp.o.d -o tests/CMakeFiles/test_sparse_walk.dir/cpp/test_sparse_walk.cpp.o -c /polytopewalk/tests/cpp/test_sparse_walk.cpp
      [27/115] /usr/bin/g++  -I/tmp/tmplbg1x2ny/build/_deps/catch2-src/src/catch2/.. -I/tmp/tmplbg1x2ny/build/generated-includes -O2 -fPIC -O3 -DNDEBUG -std=gnu++14 -ffile-prefix-map=/tmp/tmplbg1x2ny/build/_deps/catch2-src=. -MD -MT _deps/catch2-build/src/CMakeFiles/Catch2.dir/catch2/reporters/catch_reporter_cumulative_base.cpp.o -MF _deps/catch2-build/src/CMakeFiles/Catch2.dir/catch2/reporters/catch_reporter_cumulative_base.cpp.o.d -o _deps/catch2-build/src/CMakeFiles/Catch2.dir/catch2/reporters/catch_reporter_cumulative_base.cpp.o -c /tmp/tmplbg1x2ny/build/_deps/catch2-src/src/catch2/reporters/catch_reporter_cumulative_base.cpp
      [28/115] /usr/bin/g++  -I/polytopewalk/src/. -I/polytopewalk/src/Eigen3::Eigen -I/tmp/tmplbg1x2ny/build/_deps/catch2-src/src/catch2/.. -I/tmp/tmplbg1x2ny/build/generated-includes -isystem /usr/include/eigen3 -O2 -fPIC -O3 -DNDEBUG -std=gnu++14 -MD -MT tests/CMakeFiles/test_init.dir/cpp/test_init.cpp.o -MF tests/CMakeFiles/test_init.dir/cpp/test_init.cpp.o.d -o tests/CMakeFiles/test_init.dir/cpp/test_init.cpp.o -c /polytopewalk/tests/cpp/test_init.cpp
      [29/115] /usr/bin/g++  -I/polytopewalk/src/. -I/polytopewalk/src/Eigen3::Eigen -I/tmp/tmplbg1x2ny/build/_deps/catch2-src/src/catch2/.. -I/tmp/tmplbg1x2ny/build/generated-includes -isystem /usr/include/eigen3 -O2 -fPIC -O3 -DNDEBUG -std=gnu++14 -MD -MT tests/CMakeFiles/test_weights.dir/cpp/test_weights.cpp.o -MF tests/CMakeFiles/test_weights.dir/cpp/test_weights.cpp.o.d -o tests/CMakeFiles/test_weights.dir/cpp/test_weights.cpp.o -c /polytopewalk/tests/cpp/test_weights.cpp
      [30/115] /usr/bin/g++  -I/tmp/tmplbg1x2ny/build/_deps/catch2-src/src/catch2/.. -I/tmp/tmplbg1x2ny/build/generated-includes -O2 -fPIC -O3 -DNDEBUG -std=gnu++14 -ffile-prefix-map=/tmp/tmplbg1x2ny/build/_deps/catch2-src=. -MD -MT _deps/catch2-build/src/CMakeFiles/Catch2.dir/catch2/reporters/catch_reporter_console.cpp.o -MF _deps/catch2-build/src/CMakeFiles/Catch2.dir/catch2/reporters/catch_reporter_console.cpp.o.d -o _deps/catch2-build/src/CMakeFiles/Catch2.dir/catch2/reporters/catch_reporter_console.cpp.o -c /tmp/tmplbg1x2ny/build/_deps/catch2-src/src/catch2/reporters/catch_reporter_console.cpp
      [31/115] /usr/bin/g++  -I/tmp/tmplbg1x2ny/build/_deps/catch2-src/src/catch2/.. -I/tmp/tmplbg1x2ny/build/generated-includes -O2 -fPIC -O3 -DNDEBUG -std=gnu++14 -ffile-prefix-map=/tmp/tmplbg1x2ny/build/_deps/catch2-src=. -MD -MT _deps/catch2-build/src/CMakeFiles/Catch2.dir/catch2/reporters/catch_reporter_registrars.cpp.o -MF _deps/catch2-build/src/CMakeFiles/Catch2.dir/catch2/reporters/catch_reporter_registrars.cpp.o.d -o _deps/catch2-build/src/CMakeFiles/Catch2.dir/catch2/reporters/catch_reporter_registrars.cpp.o -c /tmp/tmplbg1x2ny/build/_deps/catch2-src/src/catch2/reporters/catch_reporter_registrars.cpp
      [32/115] /usr/bin/g++  -I/polytopewalk/src/. -I/polytopewalk/src/Eigen3::Eigen -I/tmp/tmplbg1x2ny/build/_deps/catch2-src/src/catch2/.. -I/tmp/tmplbg1x2ny/build/generated-includes -isystem /usr/include/eigen3 -O2 -fPIC -O3 -DNDEBUG -std=gnu++14 -MD -MT tests/CMakeFiles/test_fr.dir/cpp/test_fr.cpp.o -MF tests/CMakeFiles/test_fr.dir/cpp/test_fr.cpp.o.d -o tests/CMakeFiles/test_fr.dir/cpp/test_fr.cpp.o -c /polytopewalk/tests/cpp/test_fr.cpp
      [33/115] /usr/bin/g++  -I/tmp/tmplbg1x2ny/build/_deps/catch2-src/src/catch2/.. -I/tmp/tmplbg1x2ny/build/generated-includes -O2 -fPIC -O3 -DNDEBUG -std=gnu++14 -ffile-prefix-map=/tmp/tmplbg1x2ny/build/_deps/catch2-src=. -MD -MT _deps/catch2-build/src/CMakeFiles/Catch2.dir/catch2/reporters/catch_reporter_junit.cpp.o -MF _deps/catch2-build/src/CMakeFiles/Catch2.dir/catch2/reporters/catch_reporter_junit.cpp.o.d -o _deps/catch2-build/src/CMakeFiles/Catch2.dir/catch2/reporters/catch_reporter_junit.cpp.o -c /tmp/tmplbg1x2ny/build/_deps/catch2-src/src/catch2/reporters/catch_reporter_junit.cpp
      [34/115] /usr/bin/g++  -I/tmp/tmplbg1x2ny/build/_deps/catch2-src/src/catch2/.. -I/tmp/tmplbg1x2ny/build/generated-includes -O2 -fPIC -O3 -DNDEBUG -std=gnu++14 -ffile-prefix-map=/tmp/tmplbg1x2ny/build/_deps/catch2-src=. -MD -MT _deps/catch2-build/src/CMakeFiles/Catch2.dir/catch2/reporters/catch_reporter_streaming_base.cpp.o -MF _deps/catch2-build/src/CMakeFiles/Catch2.dir/catch2/reporters/catch_reporter_streaming_base.cpp.o.d -o _deps/catch2-build/src/CMakeFiles/Catch2.dir/catch2/reporters/catch_reporter_streaming_base.cpp.o -c /tmp/tmplbg1x2ny/build/_deps/catch2-src/src/catch2/reporters/catch_reporter_streaming_base.cpp
      [35/115] /usr/bin/g++  -I/tmp/tmplbg1x2ny/build/_deps/catch2-src/src/catch2/.. -I/tmp/tmplbg1x2ny/build/generated-includes -O2 -fPIC -O3 -DNDEBUG -std=gnu++14 -ffile-prefix-map=/tmp/tmplbg1x2ny/build/_deps/catch2-src=. -MD -MT _deps/catch2-build/src/CMakeFiles/Catch2.dir/catch2/reporters/catch_reporter_multi.cpp.o -MF _deps/catch2-build/src/CMakeFiles/Catch2.dir/catch2/reporters/catch_reporter_multi.cpp.o.d -o _deps/catch2-build/src/CMakeFiles/Catch2.dir/catch2/reporters/catch_reporter_multi.cpp.o -c /tmp/tmplbg1x2ny/build/_deps/catch2-src/src/catch2/reporters/catch_reporter_multi.cpp
      [36/115] /usr/bin/g++  -I/tmp/tmplbg1x2ny/build/_deps/catch2-src/src/catch2/.. -I/tmp/tmplbg1x2ny/build/generated-includes -O2 -fPIC -O3 -DNDEBUG -std=gnu++14 -ffile-prefix-map=/tmp/tmplbg1x2ny/build/_deps/catch2-src=. -MD -MT _deps/catch2-build/src/CMakeFiles/Catch2.dir/catch2/catch_approx.cpp.o -MF _deps/catch2-build/src/CMakeFiles/Catch2.dir/catch2/catch_approx.cpp.o.d -o _deps/catch2-build/src/CMakeFiles/Catch2.dir/catch2/catch_approx.cpp.o -c /tmp/tmplbg1x2ny/build/_deps/catch2-src/src/catch2/catch_approx.cpp
      [37/115] /usr/bin/g++  -I/tmp/tmplbg1x2ny/build/_deps/catch2-src/src/catch2/.. -I/tmp/tmplbg1x2ny/build/generated-includes -O2 -fPIC -O3 -DNDEBUG -std=gnu++14 -ffile-prefix-map=/tmp/tmplbg1x2ny/build/_deps/catch2-src=. -MD -MT _deps/catch2-build/src/CMakeFiles/Catch2.dir/catch2/reporters/catch_reporter_teamcity.cpp.o -MF _deps/catch2-build/src/CMakeFiles/Catch2.dir/catch2/reporters/catch_reporter_teamcity.cpp.o.d -o _deps/catch2-build/src/CMakeFiles/Catch2.dir/catch2/reporters/catch_reporter_teamcity.cpp.o -c /tmp/tmplbg1x2ny/build/_deps/catch2-src/src/catch2/reporters/catch_reporter_teamcity.cpp
[…]
