### 2025-05-09 ethz-randomwalk/polytopewalk#11 @matt-graham
Suggested improvements to JOSS paper text
Raising as part of openjournals/joss-reviews#7957

Some comments and suggestions on JOSS paper text, organized by corresponding top-level sections in paper:

## Summary

- The summary is intended to be readable by 'a diverse, non-specialist audience'. Currently there are several terms used without clarification which I think would not necessarily be familiar to a non-specialist audience - _polytope_,  _MCMC_, _facial reduction_. The last in that list (_facial reduction_) I think is probably okay to keep as it is an example which will presumably be recognisable to some readers but is not essential to understand the summary. I would say giving a brief high-level description of what a polytope is and what Markov chain Monte Carlo methods do would however be useful. These terms are both defined in the _Statement of Need_ section, but I would say it would be better to have brief natural language descriptions of the concepts in the summary to make it self contained. For example 'Markov chain Monte Carlo methods are a family of algorithms for generating approximate samples from a target probability distribution'.
- The phrase 'novel implementation preprocessing algorithms such as facial reduction and initialization' is a bit difficult to parse. Possibly 'novel implementations of preprocessing algorithms such as facial reduction and initialization' would be better?
- MCMC acronym needs to be expanded in first usage.
- Specifying repository URL is somewhat redundant as it is included automatically in sidebar of rendered JOSS paper.

## Statement of need

- 'We develop novel interior-point-method-based MCMC algorithms optimized for large and sparse constrained polytopes' - it's unclear to me what 'we develop' means in this context. That the authors have developed such algorithms? That they have implemented them in the PolytopeWalk package? If the former I think this is not strictly relevant to the JOSS paper which should be focused on the software package, not the underlying methods, though if there are novel methods developed by the authors which are implemented, referencing a paper in which these methods are described would be useful! If the intention is to indicate that the package implements these algorithms then I think 'implement' would be a more appropriate verb here than 'develop'.
- Ideally should add references to [Metropolis et al. (1953)](https://doi.org/10.1063%2F1.1699114) and [Hastings (1970)](https://doi.org/10.1093%2Fbiomet%2F57.1.97) for 'Metropolis-Hastings accept-reject step'.
- ' Specific algorithms include the Dikin Walk, Vaidya Walk, John Walk, and Lee Sidford Walk.' - this sentence seems a bit redundant given algorithms are listed in table in _Random Walk Algorithms_ section later, and it feels a bit odd to have them listed in text here without accompanying references (which are already included in table).
- 'in both the sparse, constrained and full-dimensional formulation' - the commas here make it appear like this is a list of three items while I think 'sparse, constrained' and 'full-dimensional' are the two groupings intended. If 'formulation' is intended to be part of both groups I think it should also be pluralized. I would say 'in both the sparse-constrained and full-dimensional formulations' would be clearer.
- 'an open-source implementation of the Facial Reduction algorithm' - it would be useful to have a reference here (or a comment that this algorithm is described in more detail in _Preprocessing algorithms_ section).

## Package overview
- The specified dependencies should have references added - [Eigen](https://eigen.tuxfamily.org/index.php?title=BibTeX), [pybind](https://pybind11.readthedocs.io/en/stable/faq.html#how-to-cite-this-project), [glpk](https://lists.gnu.org/archive/html/help-glpk/2009-04/msg00063.html). Given the Python package also depends on [NumPy](https://numpy.org/citing-numpy/) and [SciPy](https://scipy.org/citing-scipy/) it would also be good to mention this and cite accordingly.

### Random Walk Algorithms
- The meaning of the $\tau$ notation in the table is unclear to me. Is this a form of asymptotic (complexity) notation? If so it might be worth clarifying what it represents as it does not appear to be a common notation - for example I can't find any mention of it on in [this section on Wikipedia](https://en.wikipedia.org/wiki/Big_O_notation#Related_asymptotic_notations).

## References
- [In-text citations should use the syntax `[@citationid]`](https://joss.readthedocs.io/en/latest/paper.html#citations) rather than `(@citationid)` if intended to be parenthesized, and `@citationid` for non-parenthesized citations. Parenthesized lists of multiple citations should be of the form `[@citationid1;citationid2]` rather than `(@citationid1) (@citationid2)`.
- Braces should be added to BibTeX for 'polyrun: A java library for sampling from the bounded convex polytopes' to retain capitalization of _Java_.
