### 2025-05-09 ethz-randomwalk/polytopewalk#12 @matt-graham
Adding instructions on how to run tests
Raising as part of openjournals/joss-reviews#7957

It would be useful to add some instructions on how to run both the Python and C++ tests to the _Testing_ section in the project README. For the Python tests this can be inferred from looking at the `tests.yml` workflow file but it would be nice to not have to dig in to there to find out how to run tests locally!
