### 2025-05-09 ethz-randomwalk/polytopewalk#10 @yuachen
I think it has to do with Catch version in CMake. We should use the latest version @bsun1220

https://github.com/ethz-randomwalk/polytopewalk/blob/55fd336c348b5f3724079534804a4d9eca2301b0/tests/CMakeLists.txt#L7
