### 2025-05-09 ethz-randomwalk/polytopewalk#8 @bsun1220
Thanks for the feedback, I have added your suggestions to the pyproject.toml!
