### 2025-05-09 45244ef
Edited equations in Readme
