+++ README.md
@@ -31,13 +31,17 @@ For each implemented algorithm, we provide the full-dimensional formulation and 
 
 In the full-dimensional formulation with dense matrix A ($n$ x $d$ matrix) and vector b ($n$ dimensional vector), we specify the following: 
 
-![equation](https://latex.codecogs.com/svg.image?\large&space;\mathcal{K}_1=\[x\in\mathbb{R}^d|Ax\le&space;b\])
+```math
+\mathcal{K}_1 = \{x \in \mathbb{R}^{d} | Ax \le b\}
+```
 
 where the polytope is specified with $n$ constraints.
 
 In the constrained formulation with sparse matrix A ($n$ x $d$ matrix) and vector b ($n$ dimensional vector), we specify the following: 
 
-![equation](https://latex.codecogs.com/svg.image?\large&space;\mathcal{K}_2=\[x\in\mathbb{R}^d|Ax=b,x\succeq_k&space;0\])
+```math
+\mathcal{K}_2 = \{x \in \mathbb{R}^{d} | Ax = b, x \succeq_k 0\}
+```
 
 where the polytope is specified with $n$ equality constraints and $k$ coordinate-wise inequality constraints. 
 
