### 2025-05-09 ethz-randomwalk/polytopewalk#12 @bsun1220
Sounds good, I have updated the README for more detailed instructions for both C++ & Python
