### 2025-05-12 ethz-randomwalk/polytopewalk#8 @matt-graham
At the moment it looks like this hasn't been added to `pyproject.toml` yet on `main` at least as last commit changing that file still showing as a few weeks ago. Are the changes on another branch or did changes on a local `main` branch not get pushed perhaps?
