### 2025-05-12 ethz-randomwalk/polytopewalk#10 @matt-graham
Thanks for fixing - just tested again locally and package is now building without issues
