### 2025-05-13 9c77d2c
initial paper edits
