+++ paper/paper.md
@@ -24,11 +24,11 @@ bibliography: paper.bib
 
 # Summary
 
-High dimensional sampling is an important computational tool in statistics and other disciplines, with applications ranging from Bayesian uncertainty quantification, metabolic modeling in biology, to volume computation. We present ``PolytopeWalk``, a new scalable Python library designed for uniform sampling over polytopes. The library provides an end-to-end solution, that includes novel implementation preprocessing algorithms such as facial reduction and initialization. Six state-of-the-art MCMC algorithms on polytopes are implemented, including the Dikin, Vaidya, and John Walk. Additionally, we introduce novel sparse constrained formulations of these algorithms, enabling efficient sampling from sparse polytopes of the form $\mathcal{K}_2 = \{x \in \mathbb{R}^d \ | \ Ax = b, x \succeq_k 0\}$. This implementation maintains sparsity in $A$, ensuring scalability to high dimensional settings $(d > 10^4)$. ``PolytopeWalk`` is available at [github.com/ethz-randomwalk/polytopewalk](github.com/ethz-randomwalk/polytopewalk) with documentation at [polytopewalk.readthedocs.io/](polytopewalk.readthedocs.io/).
+High dimensional sampling is an important computational tool in statistics, with applications in metabolic modeling in biology, volume computation, and fast randomized algorithms. We present ``PolytopeWalk``, a scalable library designed for sampling from a uniform distribution over polytopes, defined as bounded geometric objects formed by linear inequalities. To solve this problem, we use Markov chain Monte Carlo (MCMC) methods, which are a family of algorithms for generating approximate samples from a target probability distribution. This library provides an end-to-end solution, that includes novel implementation of preprocessing algorithms such as facial reduction and initialization. Six state-of-the-art MCMC algorithms are implemented, including the Dikin, Vaidya, and John Walk. Additionally, we introduce novel sparse constrained formulations of these algorithms, enabling efficient sampling from sparse polytopes of the form $\mathcal{K}_2 = \{x \in \mathbb{R}^d \ | \ Ax = b, x \succeq_k 0\}$. This implementation maintains sparsity in $A$, ensuring scalability to high dimensional settings $(d > 10^4)$.
 
 # Statement of Need
 
-High dimensional sampling is a fundamental problem in many computational disciplines such as statistics, probability, and operation research. For example, sampling is applied in portfolio optimization (@DBLP:journals/corr/abs-1803-05861), metabolic networks in biology (@COBRA) and volume approximation over convex shapes (@Simonovits2003). Markov chain Monte Carlo (MCMC) sampling algorithms offer a natural and scalable solution to this problem. These algorithms construct a Markov chain whose stationary distribution matches the target distribution. By running the chain for a large number of steps to ensure mixing, MCMC algorithms can efficiently generate approximately independent samples close to the target distribution, while not suffering from the curse of dimension issues.
[…]
