### 2025-05-13 ethz-randomwalk/polytopewalk#11 @bsun1220
I adjusted the paper text based on your suggestions:

- For the summary, I included definitions of both MCMC and polytopes, expanded MCMC, and fixed the typo you mentioned
- For the statement of need, I replaced 'developed' with 'implemented', added the references for Metropolis-Hastings, removed the redundant sentence, fixed the typos, and added reference for facial reduction
- I added in references for the packages that I did not cite
- I fixed the citations (replacing them with square brackets) and fixed the Java capitalization error
- I replaced the tau notation with big O. For reference, I based my notation from other papers which look at both sampling time cost (using big O) and mixing time cost (using tau but it means the same thing as big O) to make it less confusing. Since we do not discuss sampling time complexity, it is not needed here.

Let me know if you have any other comments on the revised text!
