### 2025-05-14 ethz-randomwalk/polytopewalk#8 @matt-graham
Thanks, closing as now completed.
