### 2025-05-14 ethz-randomwalk/polytopewalk#11 @matt-graham
Thanks for the updates - there are a couple of minor formatting issues with the references after the changes:

- In the Hastings (1970) reference the title _Monte carlo sampling methods using markov chains and their applications._ should have _Carlo_ and _Markov_ capitalized
- In the _Statement of Need_ section `[@Metropolis1953] [@10.1093/biomet/57.1.97]` should be `[@Metropolis1953; 10.1093/biomet/57.1.97]`
- In the _Preprocessing Algorithms_ section `[@drusvyatskiy2017many][@im2023revisiting]` should be `[@drusvyatskiy2017many; im2023revisiting]` (currently not formatting correctly in rendered PDF)
