+++ paper/paper.bib
@@ -11,7 +11,7 @@
 
 @article{10.1093/biomet/57.1.97,
     author = {Hastings, W. K.},
-    title = {Monte Carlo sampling methods using Markov chains and their applications},
+    title = {{Monte Carlo sampling methods using Markov chains and their applications}},
     journal = {Biometrika},
     volume = {57},
     number = {1},
