### 2025-06-23 ethz-randomwalk/polytopewalk#18 @martinmodrak
No checks that `k` provided to `denseFullWalkRun` / `FacialReduction.reduce` is correct
A simple reproduction based on the example:

```python
import numpy as np
from scipy.sparse import csr_matrix, lil_matrix, csr_array
from polytopewalk.dense import DikinWalk, DenseCenter
from polytopewalk import FacialReduction, denseFullWalkRun

def generate_simplex(d):
    return np.array([1/d] * d), np.array([[1] * d]), np.array([1]), d, 'simplex'

fr = FacialReduction()
_, A, b, k, name = generate_simplex(5)

polytope = fr.reduce(A, b, k, sparse = False)
dense_A = polytope.dense_A
dense_b = polytope.dense_b

dc = DenseCenter()
init = dc.getInitialPoint(dense_A, dense_b)

dikin = DikinWalk(r = 0.9, thin = 1)
dikin_res = denseFullWalkRun(dense_A, dense_b, k, 1_000, dikin, fr, dc, burn = 100)
```

I get:

```
corrupted size vs. prev_size
Aborted (core dumped)
```

the problem is that I am passing the already reduced matrices, but keeping the original `k`. This should signal an error.
when `k` is set too low, I get:

```
glp_add_rows: nrs = 0; invalid number of rows
Error detected in file glpapi01.c at line 248
Aborted (core dumped)
```

More generally, I don't see a use case for user providing `k` that differs from the dimensions implied by the model matrix, so I don't think `k` should be an argument to the methods at all and should be taken internally from the matrix.
