### 2025-06-23 ethz-randomwalk/polytopewalk#18 @martinmodrak
Similarly, providing larger `k` to `FacialReduction` crashes - once again, there seems to be no reason to give `k` to `reduce()` at all.

```
import numpy as np
from scipy.sparse import csr_matrix, lil_matrix, csr_array
from polytopewalk.dense import DikinWalk, DenseCenter
from polytopewalk import FacialReduction, denseFullWalkRun

def generate_simplex(d):
    return np.array([1/d] * d), np.array([[1] * d]), np.array([1]), d, 'simplex'

fr = FacialReduction()
_, A, b, k, name = generate_simplex(5)

polytope = fr.reduce(A, b, k +1, sparse = False)
```

output:

```
double free or corruption (out)
Aborted (core dumped)
```
