### 2025-06-23 ethz-randomwalk/polytopewalk#20 @martinmodrak
No way to test correctness of the MCMC output
Currently, there are IMHO reasonable tests for the facial reduction and center initialization code. But for the MCMC code, it is only checked that it runs, not that the result is in any way correct.

Automated tests for MCMC are kind of hard to build due to the stochastic nature, but the example notebooks could be extended a little bit to at least show that means, variances and covariances and estimate of some more complex integrals involving all components (say weighted sum of squares of components) for a few examples are close enough either known analytical solutions or results obtained with reliable samplers.

E.g. for the simplex, we know the moments and can obtain high quality samples with a Dirichlet distribution RNG. Presumably, there might be other simple examples used throughout the package also have known results / known good RNGs.
