### 2025-06-23 ethz-randomwalk/polytopewalk#20 @martinmodrak
Additionally, one could/should also test convergence via running multiple chains and computing rhat/ess via Arviz
