### 2025-06-23 ethz-randomwalk/polytopewalk#22 @martinmodrak
The examples do not converge with the provided settings
The examples get too few samples to have the chains converge. For example, starting of `sparse_walks.ipynb` I add:

```python
import arviz

nchains = 4
niter = 4000
burn = 0
walks_all = np.empty([nchains, niter, k])
for c in range(nchains):
  walks_all[c, :] = sparseFullWalkRun(A, b, k, niter, sparse_dikin, fr, sc, burn = burn)

# Test with Arviz
walks_ds = arviz.convert_to_dataset(walks_all)
arviz.rhat(walks_ds)
```

For the 5D simplex I obtain split rhats similar to

```
        1.045 1.008 1.014 1.028 1.014
```

Where values above 1.01 indicate that the chains have not reached (univariate) stationary distribution.

I start to reliably get good diagnostics when running ~10 000 iterations with ~10 000 burnin. Presumably the examples should reflect that.

When I instead generate the polytope with say `generate_birkhoff(36)` I start getting convergence with ~ thin = 10, 10K iterations and 1K burnin.
