### 2025-06-23 ethz-randomwalk/polytopewalk#23 @martinmodrak
Suspicious handling of RNG state
Whenever `generateGaussianRV`, `generateSample` (and likely some other methods) are called you generate a new random number generator (see https://github.com/ethz-randomwalk/polytopewalk/blob/c6ff026acff73c09c678d8394fa327778b0a8ed5/src/dense/BarrierWalk.cpp#L9C5-L10C23) and use a `random_device` instance to initialize it:

```cpp
   random_device rd;
    mt19937 gen(rd());
```

This is problematic for two reasons:

1) It may compromise MCMC on some machines: From the [docs](https://en.cppreference.com/w/cpp/numeric/random/random_device.html) - emphasis mine:

> std::random_device may be implemented in terms of an implementation-defined pseudo-random number engine if a non-deterministic source (e.g. a hardware device) is not available to the implementation. In this case each [std::random_device](https://en.cppreference.com/w/cpp/numeric/random/random_device.html#top) object **may generate the same number sequence**.

2) If `random_device` tries to access the global entropy pool, MCMC is probably fine, but this can be a costly operation (potentially, the construction of the Mersenne Twister from the seed is also a small cost).

It would be much safer (and more performant) to create the RNG once per MCMC chain and pass a reference to it to the methods. The user-facing methods should then expose a `seed` argument to allow for reproducible results and `random_device` would only be used if the seed is not provided.
