### 2025-06-23 #7957 @martinmodrak
An update on my review. This is overall a nice package. I've encountered a bunch of minor issues and made some suggestions for API changes, none of which are critical for acceptance. The only issues I now see as important are:

- https://github.com/ethz-randomwalk/polytopewalk/issues/23 (IMHO problematic handling of RNG state)
- https://github.com/ethz-randomwalk/polytopewalk/issues/20 (missing steps to verify MCMC correctness at least for simple problems - as execpted by the "Automated test" checklist item)
- https://github.com/ethz-randomwalk/polytopewalk/issues/18 (a low-level crash for incorrect parameters)
- https://github.com/ethz-randomwalk/polytopewalk/issues/22 (examples should increase the lenghts of the MCMC chains and burnin, because under the current settings they don't converge to the stationary distribution)

I am open to discussion if the authors/editors believe those are also not critical.

My computer is currently processing some larger problems with the package to let me assess if the performance scales to "polytopes with dimensions exceeding 10^4" as claimed in the manuscript. I expect the chains to finish within days for this claim to hold.
