### 2025-06-23 b18ce79
Added package dependency to README
