### 2025-06-23 ethz-randomwalk/polytopewalk#22 @bsun1220
Just updated the examples section to reflect convergence. Let me know what you think!
