### 2025-06-24 624111a
resolved k errors
