+++ src/sparse/SparseBallWalk.cpp
@@ -8,6 +8,9 @@ MatrixXd SparseBallWalk::generateCompleteWalk(
     int k, 
     int burn = 0
 ){
+    if (k < 0 || k > A.cols()) {
+        throw std::invalid_argument("Parameter k must be between 0 and the number of columns in A.");
+    }
     MatrixXd results = MatrixXd::Zero(num_steps, A.cols());
 
     SparseLU<SparseMatrixXd> A_solver (A * A.transpose());
