+++ tests/cpp/test_dense_walk.cpp
@@ -77,41 +77,41 @@ TEST_CASE( "Test All Dense Combinations", "[require]" ){
     DenseCenter dc;
     FacialReduction fr; 
 
-    MatrixXd walk_res = denseFullWalkRun(simplex.A, simplex.b, simplex.k, 100, &john, &fr, &dc);
-    walk_res = denseFullWalkRun(simplex.A, simplex.b, simplex.k, 100, &dikinls, &fr, &dc);
+    MatrixXd walk_res = denseFullWalkRun(100, simplex.A, simplex.b, simplex.k, &john, &fr, &dc);
+    walk_res = denseFullWalkRun(100, simplex.A, simplex.b, simplex.k, &dikinls, &fr, &dc);
     REQUIRE(walk_res.rows() == 100);
-    walk_res = denseFullWalkRun(simplex.A, simplex.b, simplex.k, 100, &vaidya, &fr, &dc);
+    walk_res = denseFullWalkRun(100, simplex.A, simplex.b, simplex.k, &vaidya, &fr, &dc);
     REQUIRE(walk_res.rows() == 100);
-    walk_res = denseFullWalkRun(simplex.A, simplex.b, simplex.k, 100, &dikin, &fr, &dc);
+    walk_res = denseFullWalkRun(100, simplex.A, simplex.b, simplex.k, &dikin, &fr, &dc);
     REQUIRE(walk_res.rows() == 100);
-    walk_res = denseFullWalkRun(simplex.A, simplex.b, simplex.k, 100, &ball, &fr, &dc);
+    walk_res = denseFullWalkRun(100, simplex.A, simplex.b, simplex.k, &ball, &fr, &dc);
     REQUIRE(walk_res.rows() == 100);
-    walk_res = denseFullWalkRun(simplex.A, simplex.b, simplex.k, 100, &hitrun, &fr, &dc);
-    REQUIRE(walk_res.rows() == 100);
-
-    walk_res = denseFullWalkRun(hc.A, hc.b, hc.k, 100, &john, &fr, &dc);
-    REQUIRE(walk_res.rows() == 100);
-    walk_res = denseFullWalkRun(hc.A, hc.b, hc.k, 100, &dikinls, &fr, &dc);
-    REQUIRE(walk_res.rows() == 100);
-    walk_res = denseFullWalkRun(hc.A, hc.b, hc.k, 100, &vaidya, &fr, &dc);
-    REQUIRE(walk_res.rows() == 100);
-    walk_res = denseFullWalkRun(hc.A, hc.b, hc.k, 100, &dikin, &fr, &dc);
-    REQUIRE(walk_res.rows() == 100);
-    walk_res = denseFullWalkRun(hc.A, hc.b, hc.k, 100, &ball, &fr, &dc);
-    REQUIRE(walk_res.rows() == 100);
-    walk_res = denseFullWalkRun(hc.A, hc.b, hc.k, 100, &hitrun, &fr, &dc);
+    walk_res = denseFullWalkRun(100, simplex.A, simplex.b, simplex.k, &hitrun, &fr, &dc);
     REQUIRE(walk_res.rows() == 100);
 
-    walk_res = denseFullWalkRun(birk.A, birk.b, birk.k, 100, &john, &fr, &dc);
+    walk_res = denseFullWalkRun(100, hc.A, hc.b, hc.k, &john, &fr, &dc);
     REQUIRE(walk_res.rows() == 100);
-    walk_res = denseFullWalkRun(birk.A, birk.b, birk.k, 100, &dikinls, &fr, &dc);
[…]
