### 2025-06-24 ethz-randomwalk/polytopewalk#18 @bsun1220
Hey @martinmodrak,

I made the following adjustments (will be pushed to pypi after the other issues have fixed).
1. I added an assertion error that 0 <= k<= column length of **A**
2. I added more documentation for denseFullWalkRun & sparseFullWalkRun

For the importance of k, I agree that most of the time, you would set k = column length of **A** (meaning all variables >= 0). When we tested on the NetLib polytopes, they all had k = column length of **A**.

However, there are cases where k != column length of **A**. Conceptually, we define the constrained form as {**A**x = **b**, x >=_k 0} meaning the last k coordinates are >= 0 and there are no restrictions on the first d - k coordinates.

If we wanted to create a 2D hypercube (all coordinates between -1 and 1) in this formulation. **A** would be of the form:

|1, 0, 1, 0, 0, 0|
|-1, 0, 0, 1, 0, 0|
|0, 1, 0, 0, 1, 0|
|0, -1, 0, 0, 0, 1|

and **b** = [1, 1, 1, 1] where the last 4 coordinates of x would be >= 0. Then, we would just take the first 2 coordinates as our samples.

By letting k not be set to column length of **A**, we ensure that for any formulation in full-dimensional form, there is a corresponding formulation in the constrained form.

Also, there have been some confusion with the denseFullWalkRun, which I added additional documentation to. denseFullWalkRun takes as inputs a polytope in the constrained formulation, does facial reduction (which can be used to convert to the full-dimensional form), run the sampler on the full-dimensional form, then converts it back to the original dimension. Therefore, the input to both the denseFullWalkRun and sparseFullWalkRun is the same.

The only difference is the execution of the MCMC sampler (there may be cases where it is just faster to run on the full-dimensional form than the constrained form). In your code, you would be passing in a polytope already in the full-dimensional form. Again, I have added to the documentation to alleviate this issue. Let me know what you think!
