+++ src/sparse/SparseBallWalk.cpp
@@ -6,7 +6,8 @@ MatrixXd SparseBallWalk::generateCompleteWalk(
     const SparseMatrixXd& A, 
     const VectorXd& b, 
     int k, 
-    int burn = 0
+    int burn = 0,
+    int seed = -1
 ){
     if (k < 0 || k > A.cols()) {
         throw std::invalid_argument("Parameter k must be between 0 and the number of columns in A.");
@@ -16,11 +17,13 @@ MatrixXd SparseBallWalk::generateCompleteWalk(
     SparseLU<SparseMatrixXd> A_solver (A * A.transpose());
     SparseMatrixXd I = SparseMatrixXd(VectorXd::Ones(A.cols()).asDiagonal());
 
+    std::mt19937 gen = initializeRNG(seed);
+
     VectorXd x = init;
     int d = A.cols() - A.rows();
     int total = (burn + num_steps) * THIN;
     for (int i = 1; i <= total; i++){
-        VectorXd rand = generateGaussianRV(A.cols()); 
+        VectorXd rand = generateGaussianRV(A.cols(), gen); 
         VectorXd z;
         z = A * rand; 
         z = rand - A.transpose() * A_solver.solve(z);
