+++ src/dense/RandomWalk.cpp
@@ -4,10 +4,8 @@ bool RandomWalk::inPolytope(const VectorXd& vec, const MatrixXd& A, const Vector
     return ((A * vec) - b).maxCoeff() <= 0;
 }
 
-VectorXd RandomWalk::generateGaussianRVNorm(const int d){
+VectorXd RandomWalk::generateGaussianRVNorm(const int d, std::mt19937& gen){
     VectorXd v(d);
-    random_device rd;
-    mt19937 gen(rd());
     normal_distribution<double> dis(0.0, 1.0);
     for(int i = 0; i < d; i++){
         v(i) = dis(gen);
@@ -15,11 +13,20 @@ VectorXd RandomWalk::generateGaussianRVNorm(const int d){
     return v/v.norm();
 }
 
-MatrixXd RandomWalk::generateCompleteWalk(const int num_steps, VectorXd& x, const MatrixXd& A, const VectorXd& b, int burn = 0){
+MatrixXd RandomWalk::generateCompleteWalk(const int num_steps, VectorXd& x, const MatrixXd& A, const VectorXd& b, int burn = 0, int seed = -1){
     cout << "oops" << endl;
     return MatrixXd::Zero(1,1);
 }
 
 void RandomWalk::printType(){
     cout << "oops" << endl;
+}
+
+std::mt19937 RandomWalk::initializeRNG(int seed) {
+    if (seed != -1) {
+        return std::mt19937(seed);
+    } else {
+        std::random_device rd;
+        return std::mt19937(rd());
+    }
 }
