+++ src/sparse/SparseBallWalk.hpp
@@ -20,6 +20,7 @@ class SparseBallWalk : public SparseRandomWalk{
          * @param b polytope vector
          * @param k k values >= 0 constraint
          * @param burn number of initial steps to cut
+         * @param seed seed for reproducibility
          * @return num_steps by d (dimension of x) matrix
          */
         MatrixXd generateCompleteWalk(
@@ -28,7 +29,8 @@ class SparseBallWalk : public SparseRandomWalk{
             const SparseMatrixXd& A, 
             const VectorXd& b, 
             int k, 
-            int burn
+            int burn,
+            int seed
             ) override;
     
     protected:
