### 2025-06-25 0c83d06
Update test.yml for macos
