### 2025-06-25 5667ec4
improved latency on dense mcmc
