+++ src/dense/RandomWalk.hpp
@@ -15,14 +15,14 @@ class RandomWalk{
         /**
          * @brief generate values from the walk
          * @param num_steps number of steps wanted to take
-         * @param x initial starting point
+         * @param init initial starting point
          * @param A polytope matrix (Ax <= b)
          * @param b polytope vector (Ax <= b)
          * @param burn number of initial steps to cut
          * @param seed seed for reproducibility
          * @return num_steps by d (dimension of x) matrix
          */
-        virtual MatrixXd generateCompleteWalk(const int num_steps, VectorXd& x, const MatrixXd& A, const VectorXd& b, int burn, int seed);
+        virtual MatrixXd generateCompleteWalk(const int num_steps, VectorXd& init, const MatrixXd& A, const VectorXd& b, int burn, int seed);
 
     protected: 
 
