### 2025-06-25 814d000
barrier walk efficiency
