### 2025-06-26 4f842a9
adjusted thin, niter, burnin
