+++ pybind11/PybindExt.cpp
@@ -6,16 +6,17 @@ namespace py = pybind11;
 template <class RandomWalkBase = RandomWalk> class PyRandomWalk : public RandomWalkBase {
 public:
     using RandomWalkBase::RandomWalkBase; // Inherit constructors
-    MatrixXd generateCompleteWalk(const int num_steps, VectorXd& init, const MatrixXd& A, const VectorXd& b, int burn, int seed) override{
+    MatrixXd generateCompleteWalk(const int niter, VectorXd& init, const MatrixXd& A, const VectorXd& b, int burnin, int thin, int seed) override{
             PYBIND11_OVERRIDE_PURE(
                 MatrixXd,
                 RandomWalkBase,
                 generateCompleteWalk,
-                num_steps,
+                niter,
                 init,
                 A,
                 b,
-                burn,
+                burnin,
+                thin, 
                 seed
             );
     }
@@ -24,16 +25,17 @@ public:
 template <class BarrierWalkBase = BarrierWalk> class PyBarrierWalk: public PyRandomWalk<BarrierWalkBase> {
 public:
     using PyRandomWalk<BarrierWalkBase>::PyRandomWalk;
-    MatrixXd generateCompleteWalk(const int num_steps, VectorXd& init, const MatrixXd& A, const VectorXd& b, int burn, int seed) override{
+    MatrixXd generateCompleteWalk(const int niter, VectorXd& init, const MatrixXd& A, const VectorXd& b, int burnin, int thin, int seed) override{
             PYBIND11_OVERRIDE(
                 MatrixXd,
                 BarrierWalkBase,
                 generateCompleteWalk,
-                num_steps,
+                niter,
                 init,
                 A,
                 b,
-                burn,
+                burnin,
+                thin, 
                 seed
             );
     }
@@ -63,12 +65,13 @@ template <class SparseRandomWalkBase = SparseRandomWalk> class PySparseRandomWal
 public:
     using SparseRandomWalkBase::SparseRandomWalkBase; // Inherit constructors
     MatrixXd generateCompleteWalk(
-        const int num_steps, 
+        const int niter, 
         const VectorXd& init, 
         const SparseMatrixXd& A, 
         const VectorXd& b, 
         int k, 
-        int burn,
+        int burnin,
+        int thin, 
         int seed
         ) override
     {
@@ -76,12 +79,13 @@ public:
                 MatrixXd,
                 SparseRandomWalkBase,
                 generateCompleteWalk,
-                num_steps,
+                niter,
                 init,
                 A,
                 b,
                 k,
-                burn,
+                burnin,
+                thin, 
                 seed
             );
     }
@@ -90,12 +94,13 @@ template <class SparseBarrierWalkBase = SparseBarrierWalk> class PySparseBarrier
 public:
     using PySparseRandomWalk<SparseBarrierWalkBase>::PySparseRandomWalk; 
     MatrixXd generateCompleteWalk(
-        const int num_steps, 
+        const int niter, 
         const VectorXd& init, 
         const SparseMatrixXd& A, 
         const VectorXd& b,
         int k,
-        int burn,
+        int burnin,
+        int thin, 
         int seed
         ) override
         {
@@ -103,12 +108,13 @@ public:
                 MatrixXd,
                 SparseBarrierWalkBase,
                 generateCompleteWalk,
-                num_steps,
+                niter,
                 init,
                 A,
                 b,
                 k,
-                burn,
+                burnin,
+                thin, 
                 seed
             );
     }
[…]
