### 2025-06-26 762a44d
updated examples/readme
