+++ tests/python/test_sparse_walk.py
@@ -51,26 +51,26 @@ class TestSparseWalk(unittest.TestCase):
         init_hc = sc.getInitialPoint(fr_res_hc.sparse_A, fr_res_hc.sparse_b, 4)
         init_birk = sc.getInitialPoint(fr_res_birk.sparse_A, fr_res_birk.sparse_b, 4)
 
-        res1 = dikin.generateCompleteWalk(100, init_simplex, fr_res_simplex.sparse_A, fr_res_simplex.sparse_b, 10, 10, seed = 5)
+        res1 = dikin.generateCompleteWalk(100, init_simplex, fr_res_simplex.sparse_A, fr_res_simplex.sparse_b, 3, 10, 10, seed = 5)
         self.assertTrue(res1.shape == (9, 3))
-        res2 = vaidya.generateCompleteWalk(100, init_simplex, fr_res_simplex.sparse_A, fr_res_simplex.sparse_b, 10, 10, seed = 5)
+        res2 = vaidya.generateCompleteWalk(100, init_simplex, fr_res_simplex.sparse_A, fr_res_simplex.sparse_b, 3, 10, 10, seed = 5)
         self.assertTrue(res2.shape == (9, 3))
-        res3 = john.generateCompleteWalk(100, init_simplex, fr_res_simplex.sparse_A, fr_res_simplex.sparse_b, 10, 10, seed = 5)
+        res3 = john.generateCompleteWalk(100, init_simplex, fr_res_simplex.sparse_A, fr_res_simplex.sparse_b, 3, 10, 10, seed = 5)
         self.assertTrue(res3.shape == (9, 3))
 
 
-        res1 = dikin.generateCompleteWalk(100, init_hc, fr_res_hc.sparse_A, fr_res_hc.sparse_b, 10, 10, seed = 5)
+        res1 = dikin.generateCompleteWalk(100, init_hc, fr_res_hc.sparse_A, fr_res_hc.sparse_b, 4, 10, 10, seed = 5)
         self.assertTrue(res1.shape[0] == 9)
-        res2 = vaidya.generateCompleteWalk(100, init_hc, fr_res_hc.sparse_A, fr_res_hc.sparse_b, 10, 10, seed = 5)
+        res2 = vaidya.generateCompleteWalk(100, init_hc, fr_res_hc.sparse_A, fr_res_hc.sparse_b, 4, 10, 10, seed = 5)
         self.assertTrue(res2.shape[0] == 9)
-        res3 = john.generateCompleteWalk(100, init_hc, fr_res_hc.sparse_A, fr_res_hc.sparse_b, 10, 10, seed = 5)
+        res3 = john.generateCompleteWalk(100, init_hc, fr_res_hc.sparse_A, fr_res_hc.sparse_b, 4, 10, 10, seed = 5)
         self.assertTrue(res3.shape[0] == 9)
 
-        res1 = dikin.generateCompleteWalk(100, init_birk, fr_res_birk.sparse_A, fr_res_birk.sparse_b, 10, 10, seed = 5)
+        res1 = dikin.generateCompleteWalk(100, init_birk, fr_res_birk.sparse_A, fr_res_birk.sparse_b, 4, 10, 10, seed = 5)
         self.assertTrue(res1.shape[0] == 9)
-        res2 = vaidya.generateCompleteWalk(100, init_birk, fr_res_birk.sparse_A, fr_res_birk.sparse_b, 10, 10, seed = 5)
+        res2 = vaidya.generateCompleteWalk(100, init_birk, fr_res_birk.sparse_A, fr_res_birk.sparse_b, 4, 10, 10, seed = 5)
         self.assertTrue(res2.shape[0] == 9)
[…]
