### 2025-06-27 ethz-randomwalk/polytopewalk#18 @martinmodrak
You are right having `k` for sparse formulation is completely sensible.

I just checked and still have a case that crashes, specifically, invoking `FacialReduction` with `k = 0` then crashes `DenseCenter.getInitialPoint`:

```python
import numpy as np
from scipy.sparse import csr_matrix, lil_matrix, csr_array
from polytopewalk.dense import DikinWalk, DenseCenter
from polytopewalk import FacialReduction

def generate_simplex(d):
    return np.array([1/d] * d), np.array([[1] * d]), np.array([1]), d, 'simplex'

fr = FacialReduction()
_, A, b, k, name = generate_simplex(5)
dikin = DikinWalk(r = 0.9)

polytope = fr.reduce(A, b, 0, sparse = False)
dense_A = polytope.dense_A
dense_b = polytope.dense_b

dc = DenseCenter()
init = dc.getInitialPoint(dense_A, dense_b)
```
