### 2025-06-27 aba133e
uniform testing, k fix, burn fix
