+++ src/sparse/SparseBarrierWalk.cpp
@@ -25,8 +25,8 @@ VectorXd SparseBarrierWalk::generateSample(
     int k,
     std::mt19937& gen
 ){
-    if (k < 0 || k > A.cols()) {
-        throw std::invalid_argument("Parameter k must be between 0 and the number of columns in A.");
+    if (k <= 0 || k > A.cols()) {
+        throw std::invalid_argument("Parameter k must be between 1 and the number of columns in A.");
     }
     SparseMatrixXd slack_inv = generateSlackInverse(x, k);
     SparseMatrixXd W = SparseMatrixXd(generateWeight(x, A, k).asDiagonal());
