+++ src/sparse/SparseBallWalk.cpp
@@ -10,8 +10,8 @@ MatrixXd SparseBallWalk::generateCompleteWalk(
     int thin = 1,
     int seed = -1
 ){
-    if (k < 0 || k > A.cols()) {
-        throw std::invalid_argument("Parameter k must be between 0 and the number of columns in A.");
+    if (k <= 0 || k > A.cols()) {
+        throw std::invalid_argument("Parameter k must be between 1 and the number of columns in A.");
     }
     if (init.rows() != A.cols() || A.rows() != b.rows() ) {
         throw std::invalid_argument("A, b, and init do not match in dimension.");
