+++ src/utils/FacialReduction.cpp
@@ -148,8 +148,8 @@ FRResult FacialReduction::entireFacialReductionStep(SparseMatrixXd& A, VectorXd&
 }
 
 FROutput FacialReduction::reduce(SparseMatrixXd A, VectorXd b, int k, bool sparse){
-    if (k < 0 || k > A.cols()) {
-        throw std::invalid_argument("Parameter k must be between 0 and the number of columns in A.");
+    if (k <= 0 || k > A.cols()) {
+        throw std::invalid_argument("Parameter k must be between 1 and the number of columns in A.");
     }
 
     int x_dim = A.cols() - k; 
