+++ requirements.txt
@@ -1,3 +1,4 @@
 numpy
 pandas
 scipy
+arviz
