### 2025-06-27 ethz-randomwalk/polytopewalk#20 @bsun1220
I added an additional examples notebook for testing convergence on a hypercube and a simplex. I also added those scripts to the test folder. Let me know if there is anything else!
