### 2025-06-27 e1e53ef
added additional k check
