### 2025-06-27 cdd6996
additional k check
