### 2025-06-28 #7957 @bsun1220
Hey @martinmodrak

We can adjust the writing to focus on the iteration cost rather than sampling time. From our previous experiments for sampling and getting a reasonable ess from sparse polytopes from 10^3-10^5 dimensions, it will usually take anywhere between 24 hours to several weeks, which may be possible for some research labs but likely intractable for a researcher without a serious compute cluster.

For testing the convergence, the MCMC sampler can also be sensitive to the R hyperparameter (too large means we have too many rejections and too small means the walker will not explore as much space). In the future, we plan on implementing methods like CHRMC which have improved mixing times over the interior-point method MCMC walks (Dikin/Vaidya/John).
