### 2025-06-28 #7957 @bsun1220
I just modified the paper to explain that the improvements to the existing MCMC algorithms come mainly from the per-iteration cost.
