### 2025-06-29 #7957 @martinmodrak
OK, good enough for me. @vissarion all my concerns have been satisfactorily resolved, I congratulate the authors on a neat package and recommend for publication in JOSS.
