### 2025-07-10 ethz-randomwalk/polytopewalk#25 @vissarion
JOSS review
This is a list of suggestions/comments for the JOSS manuscript:

- line 44: It would be useful to also report the number of steps needed to generate on sample
- line 57: typo, "an an"
- In line 11 the authors refer to 6 MCMC implemented algorithms and in line 54 to four algorithms and one more in the sequel of the paragraph. Also in line 72 there is a reference for 2 preprocessing algorithms. I think that could be written more clearly.
- line 85-86 volesti C++ library has a recent JOSS related paper https://doi.org/10.21105/joss.07886, what it is currently cited in about the R interface although not mentioned in the text (there is also a CRAN package https://cran.r-project.org/package=volesti). dingo, the python interface is published in https://doi.org/10.1093/bioadv/vbae037
- A relevant package that I think should be mentioned is https://dl.acm.org/doi/10.5555/3600270.3602567
- It could be very useful for the reader of this manuscript to also read the arxiv paper https://arxiv.org/pdf/2412.06629 where there is more detailed description of algorithms, experiments etc. Thus, I think it makes sense to cite it.
- For the reference in lines 95-97 there is a journal version https://doi.org/10.1016/j.comgeo.2022.101916
- Some DOI's are missing from the reference list e.g. in line 118, 130; I found those https://doi.org/10.1016/j.ejor.2023.03.021, https://doi.org/10.1007/s10107-003-0447-x
- line 99: r should be capitalized
