### 2025-07-10 #7957 @vissarion
@bsun1220 I have created an issue with some comments related to the manuscript https://github.com/ethz-randomwalk/polytopewalk/issues/25
