### 2025-07-10 a6a4119
initial paper edits
