+++ paper/paper.bib
@@ -1,14 +1,3 @@
-
-@Article{chow:68,
-  author = 	 {C. K. Chow and C. N. Liu},
-  title = 	 {Approximating discrete probability distributions with dependence trees},
-  journal = 	 {IEEE Transactions on Information Theory},
-  year = 	 {1968},
-  volume = 	 {IT-14},
-  number = 	 {3},
-  pages = 	 {462--467}}
-
-
 @article{10.1093/biomet/57.1.97,
     author = {Hastings, W. K.},
     title = {{Monte Carlo sampling methods using Markov chains and their applications}},
@@ -113,15 +102,6 @@
 }
 
 
-@Book{pearl:88,
-  author = 	 {Judea Pearl},
-  title = 	 {Probabilistic {R}easoning in {I}ntelligent {S}ystems: 
-		  {N}etworks of {P}lausible {I}nference},
-  publisher = 	 {Morgan Kaufman Publishers},
-  year = 	 {1988},
-  address = 	 {San Mateo, CA}
-}
-
 @article{DBLP:journals/corr/abs-1911-05656,
   author       = {Aditi Laddha and
                   Yin Tat Lee and
@@ -138,22 +118,20 @@
   bibsource    = {dblp computer science bibliography, https://dblp.org}
 }
 
-@article{DBLP:journals/corr/abs-2202-01908,
-  author       = {Yunbum Kook and
-                  Yin Tat Lee and
-                  Ruoqi Shen and
-                  Santosh S. Vempala},
-  title        = {Sampling with Riemannian Hamiltonian Monte Carlo in a Constrained
-                  Space},
-  journal      = {CoRR},
-  volume       = {abs/2202.01908},
-  year         = {2022},
-  url          = {https://arxiv.org/abs/2202.01908},
-  eprinttype    = {arXiv},
-  eprint       = {2202.01908},
-  timestamp    = {Wed, 09 Feb 2022 15:43:34 +0100},
-  biburl       = {https://dblp.org/rec/journals/corr/abs-2202-01908.bib},
-  bibsource    = {dblp computer science bibliography, https://dblp.org}
+@inproceedings{10.5555/3600270.3602567,
+author = {Kook, Yunbum and Lee, Yin Tat and Shen, Ruoqi and Vempala, Santosh S.},
+title = {Sampling with riemannian Hamiltonian Monte Carlo in a constrained space},
+year = {2022},
+isbn = {9781713871088},
+publisher = {Curran Associates Inc.},
+address = {Red Hook, NY, USA},
+abstract = {We demonstrate for the first time that ill-conditioned, non-smooth, constrained distributions in very high dimension, upwards of 100,000, can be sampled efficiently in practice. Our algorithm incorporates constraints into the Riemannian version of Hamiltonian Monte Carlo and maintains sparsity. This allows us to achieve a mixing rate independent of condition numbers.On benchmark data sets from systems biology and linear programming, our algorithm outperforms existing packages by orders of magnitude. In particular, we achieve a 1,000-fold speed-up for sampling from the largest published human metabolic network (RECON3D). Our package has been incorporated into the COBRA toolbox.},
+booktitle = {Proceedings of the 36th International Conference on Neural Information Processing Systems},
+articleno = {2297},
+numpages = {13},
+location = {New Orleans, LA, USA},
+series = {NIPS '22}
+doi = {10.5555/3600270.3602567}
 }
 
 @article{JMLR:v19:18-158,
@@ -164,25 +142,29 @@
   volume  = {19},
   number  = {55},
   pages   = {1--86},
-  url     = {http://jmlr.org/papers/v19/18-158.html}
[…]
