+++ paper/paper.bib
@@ -130,7 +130,7 @@ booktitle = {Proceedings of the 36th International Conference on Neural Informat
 articleno = {2297},
 numpages = {13},
 location = {New Orleans, LA, USA},
-series = {NIPS '22}
+series = {NIPS '22},
 doi = {10.5555/3600270.3602567}
 }
 
