+++ paper/paper.md
@@ -47,11 +47,11 @@ This package focuses on sampling from a uniform distribution over a user-specifi
     where $\mathcal{K}_2$ is specified via $n$ equality constraints and $k$ coordinate inequality constraints. 
 \end{enumerate}
 
-Large polytopes with sparse constraints are common in many applications [@10.5555/3600270.3602567]. The largest human metabolic network RECON3D is modeled as a $13543$-dimensional sparse polytope [@10.1093/nar/gkv1049]. Moreover, linear programming datasets from `NetLib` are naturally in the constrained form, where $A$ matrix is sparse. These applications motivate the need for MCMC algorithms that leverage $\mathcal{K}_2$ form. We implement novel interior-point-method-based MCMC algorithms optimized for large and sparse constrained polytopes. By exploiting sparsity, our algorithms scale well in per-iteration cost as a function of increasing dimension. Using the Dikin Walk, we can perform over 300 steps per second for a 10^4 dimensional simplex. For reference, a heuristic for generating 1 independent sample is the 100 steps times the mixing time. 
+Large polytopes with sparse constraints are common in many applications [@10.5555/3600270.3602567]. The largest human metabolic network RECON3D is modeled as a $13543$-dimensional sparse polytope [@10.1093/nar/gkv1049]. Moreover, linear programming datasets from `NetLib` are naturally in the constrained form, where $A$ matrix is sparse. These applications motivate the need for MCMC algorithms that leverage $\mathcal{K}_2$ form. We implement novel interior-point-method-based MCMC algorithms optimized for large and sparse constrained polytopes. By exploiting sparsity, our algorithms scale well in per-iteration cost as a function of increasing dimension. Using the Dikin Walk, we can perform over 300 steps per second for a 10^4 dimensional simplex. For reference, a heuristic for generating 1  sample is the 100 steps times the mixing time. 
 
 Interior-point-method-based MCMC sampling algorithms on a polytope are modifications of the Ball Walk [@vempala2005], incorporating key concepts from interior-point methods in optimization. These algorithms operate in two primary steps. First, the algorithm generates a proposal distribution whose covariance matrix is state-dependent and equal to the inverse of the Hessian matrix of a specified barrier function, capturing the local geometry of the polytope. Second, the algorithm employs the Metropolis-Hastings accept-reject step to ensure that its stationary distribution is uniform on the polytope [@Metropolis1953; @10.1093/biomet/57.1.97]. Using a state-dependent proposal distribution that adapts to the polytope's local geometry, these MCMC algorithms achieve an improved mixing rate.
 
-In ``PolytopeWalk``, we implement 4 interior-point-method-based MCMC sampling algorithms and 2 standard random walk algorithms (6 total) in both the sparse-constrained and full-dimensional formulation. ``PolytopeWalk`` makes meaningful strides in the open-source development of MCMC, speeding up calculations for sparse high-dimensional sampling. Finally, we provide an open-source implementation of the Facial Reduction algorithm, described in detail in the Preprocessing Algorithms section. Additional technical details can be found in a separate paper [@sun2024polytopewalksparsemcmcsampling].
+In ``PolytopeWalk``, we implement 4 barrier MCMC sampling algorithms and 2 standard random walk algorithms (6 total) in the sparse-constrained and full-dimensional formulation. ``PolytopeWalk`` makes meaningful strides in the open-source development of MCMC, speeding up calculations for sparse high-dimensional sampling. Finally, we provide an implementation of the Facial Reduction algorithm, described in detail in the Preprocessing Algorithms section. Additional technical details can be found in a separate paper [@sun2024polytopewalksparsemcmcsampling].
 
 # Package Overview
 
