+++ paper/paper.bib
@@ -19,7 +19,7 @@
    author = {Wenzel Jakob and Jason Rhinelander and Dean Moldovan},
    year = {2017},
    note = {https://github.com/pybind/pybind11},
-   title = {pybind11 -- Seamless operability between C++11 and Python}
+   title = {{pybind11 -- Seamless operability between C++11 and Python}}
 }
  @MISC{eigenweb,
   author = {Ga\"{e}l Guennebaud and Beno\^{i}t Jacob and others},
@@ -120,7 +120,7 @@
 
 @inproceedings{10.5555/3600270.3602567,
 author = {Kook, Yunbum and Lee, Yin Tat and Shen, Ruoqi and Vempala, Santosh S.},
-title = {Sampling with riemannian Hamiltonian Monte Carlo in a constrained space},
+title = {Sampling with {riemannian Hamiltonian Monte Carlo} in a constrained space},
 year = {2022},
 isbn = {9781713871088},
 publisher = {Curran Associates Inc.},
@@ -136,7 +136,7 @@ doi = {10.5555/3600270.3602567}
 
 @article{JMLR:v19:18-158,
   author  = {Yuansi Chen and Raaz Dwivedi and Martin J. Wainwright and Bin Yu},
-  title   = {Fast MCMC Sampling Algorithms on Polytopes},
+  title   = {Fast {MCMC} Sampling Algorithms on Polytopes},
   journal = {Journal of Machine Learning Research},
   year    = {2018},
   volume  = {19},
@@ -164,7 +164,7 @@ author = {Dmitriy Drusvyatskiy and Henry Wolkowicz}
 @article{DBLP:journals/corr/SachdevaV15,
   author       = {Sushant Sachdeva and
                   Nisheeth K. Vishnoi},
-  title        = {The Mixing Time of the Dikin Walk in a Polytope - A Simple Proof},
+  title        = {{The Mixing Time of the Dikin Walk in a Polytope - A Simple Proof}},
   journal      = {CoRR},
   volume       = {abs/1508.01977},
   year         = {2015},
@@ -260,7 +260,7 @@ abstract = {We examine volume computation of general-dimensional polytopes and m
 }
 
 @misc{sun2024polytopewalksparsemcmcsampling,
-      title={PolytopeWalk: Sparse MCMC Sampling over Polytopes}, 
+      title={{PolytopeWalk: Sparse MCMC Sampling over Polytopes}}, 
       author={Benny Sun and Yuansi Chen},
       year={2024},
       eprint={2412.06629},
@@ -271,7 +271,7 @@ abstract = {We examine volume computation of general-dimensional polytopes and m
 
 @article{10.1093/bioadv/vbae037,
     author = {Chalkis, Apostolos and Fisikopoulos, Vissarion and Tsigaridas, Elias and Zafeiropoulos, Haris},
-    title = {dingo: a Python package for metabolic flux sampling},
+    title = {dingo: a {Python} package for metabolic flux sampling},
     journal = {Bioinformatics Advances},
     volume = {4},
     number = {1},
@@ -286,9 +286,9 @@ abstract = {We examine volume computation of general-dimensional polytopes and m
 }
 
 
-@article{Chalkis2025, doi = {10.21105/joss.07886}, url = {https://doi.org/10.21105/joss.07886}, year = {2025}, publisher = {The Open Journal}, volume = {10}, number = {108}, pages = {7886}, author = {Chalkis, Apostolos and Fisikopoulos, Vissarion and Papachristou, Marios and Tsigaridas, Elias}, title = {volesti: A C++ library for sampling and volume computation on convex bodies}, journal = {Journal of Open Source Software} }
[…]
