### 2025-07-10 ethz-randomwalk/polytopewalk#25 @bsun1220
Hey @vissarion,

I have made the edits following your suggestions. Let me know if there are any issues!

Benny
