### 2025-07-11 ethz-randomwalk/polytopewalk#25 @vissarion
Thanks @bsun1220! I have two more comments here.

- The statement in lines 88-89 is not exact since it seems that there is an open source package related to that paper, see https://github.com/ConstrainedSampler/PolytopeSamplerMatlab
- The doi in in line 137 is not working.
