+++ paper/paper.bib
@@ -120,7 +120,7 @@
 
 @inproceedings{10.5555/3600270.3602567,
 author = {Kook, Yunbum and Lee, Yin Tat and Shen, Ruoqi and Vempala, Santosh S.},
-title = {Sampling with {riemannian Hamiltonian Monte Carlo} in a constrained space},
+title = {Sampling with riemannian Hamiltonian Monte Carlo in a constrained space},
 year = {2022},
 isbn = {9781713871088},
 publisher = {Curran Associates Inc.},
@@ -130,9 +130,9 @@ booktitle = {Proceedings of the 36th International Conference on Neural Informat
 articleno = {2297},
 numpages = {13},
 location = {New Orleans, LA, USA},
-series = {NIPS '22},
-doi = {10.5555/3600270.3602567}
+series = {NIPS '22}
 }
+
 
 @article{JMLR:v19:18-158,
   author  = {Yuansi Chen and Raaz Dwivedi and Martin J. Wainwright and Bin Yu},
