+++ paper/paper.md
@@ -83,7 +83,7 @@ Mixing times refer to the required number of steps to converge to stationary dis
 ## Package Comparison
 
 | Feature                  | ``PolytopeWalk`` | `Volesti` | `PolytopeSampler` | `Polyrun` |
-|--------------------------|:-------------:|:--------:|:------:|:--------:|
+|--------------------------|:-------------:|:--------:|:-----------------:|:--------:|
 | Constrained Formulation  | $${\color{green}Y}$$ | $${\color{red}N}$$ | $${\color{green}Y}$$ | $${\color{green}Y}$$ |
 | Sparse Friendly          | $${\color{green}Y}$$ | $${\color{red}N}$$ | $${\color{green}Y}$$ | $${\color{red}N}$$ |
 | C++ Implementation       | $${\color{green}Y}$$ | $${\color{green}Y}$$ | $${\color{green}Y}$$ | $${\color{red}N}$$ |
@@ -94,7 +94,7 @@ Mixing times refer to the required number of steps to converge to stationary dis
 | Lee-Sidford Walk         | $${\color{green}Y}$$ | $${\color{red}N}$$ | $${\color{red}N}$$ | $${\color{red}N}$$ |
 
 
-Table II contrasts the features of ``PolytopeWalk`` with `Volesti` [@Chalkis2025], `PoltopeSampler`[@10.5555/3600270.3602567], and `Polyrun` [@CIOMEK2021100659]. `Volesti` is implemented in C++ with some of its code represented in the Python library `Dingo` [@10.1093/bioadv/vbae037]. `PolytopeSampler` only works on Matlab and `Polyrun` on Java. Thus, ``PolytopeWalk`` adds additional features and novelties not found in other MCMC sampling packages.
+Table II contrasts the features of ``PolytopeWalk`` with `Volesti` [@Chalkis2025], `PoltopeSampler` [@10.5555/3600270.3602567], and `Polyrun` [@CIOMEK2021100659]. `Volesti` is implemented in C++ with some of its code represented in the Python library `Dingo` [@10.1093/bioadv/vbae037]. `PolytopeSampler` only works on Matlab and `Polyrun` on Java. Thus, ``PolytopeWalk`` adds additional features and novelties not found in other MCMC sampling packages.
 
 # Acknowledgements
 
