### 2025-07-11 ethz-randomwalk/polytopewalk#25 @bsun1220
Sounds good, I added it in the comparison table and I removed the error in line 137.
