### 2025-07-15 ethz-randomwalk/polytopewalk#25 @vissarion
Thanks for addressing the comments!

There is a typo introduced in line 86 `PoltopeSampler -> PolytopeSampler`. After fixing I am OK to close the issue.
