### 2025-07-15 #7957 @vissarion
Thanks @bsun1220! Only a minor typo is left as a comment to the issue, after fixing I am OK from my side.

This is a highly interesting and impactful piece of work that significantly contributes to the ongoing development of practical MCMC algorithms for sampling constraint spaces. Many thanks!
