### 2025-07-15 #7957 @vissarion
After fixing the issue https://github.com/ethz-randomwalk/polytopewalk/issues/25 please follow the instructions below.

When a submission is ready to be accepted, we ask that the authors issue a new tagged release of the software (if changed), and archive it (see [this guide](https://docs.github.com/en/repositories/archiving-a-github-repository/referencing-and-citing-content)). Please do this and post the version number and archive DOI here.
Please make sure that the author names, affiliations as well as the title of the archive (e.g. zenodo) is exactly the same with the submission.
