### 2025-07-15 ethz-randomwalk/polytopewalk#25 @bsun1220
Fixed the issue!
