### 2025-07-17 #7957 @vissarion
Thanks @bsun1220 please in zenodo page:
1. fix the title  (from "ethz-randomwalk/polytopewalk: v1.1.0" to "PolytopeWalk: Sparse MCMC Sampling over Polytopes")
2. add affiliations in authors and
3. remove the third author (who does not appear in the paper).
