+++ paper/paper.bib
@@ -115,7 +115,8 @@
   eprint       = {1911.05656},
   timestamp    = {Mon, 02 Dec 2019 13:44:01 +0100},
   biburl       = {https://dblp.org/rec/journals/corr/abs-1911-05656.bib},
-  bibsource    = {dblp computer science bibliography, https://dblp.org}
+  bibsource    = {dblp computer science bibliography, https://dblp.org},
+  doi          = {10.1145/3357713.3384272}
 }
 
 @inproceedings{10.5555/3600270.3602567,
@@ -142,8 +143,7 @@ series = {NIPS '22}
   volume  = {19},
   number  = {55},
   pages   = {1--86},
-  url     = {http://jmlr.org/papers/v19/18-158.html},
-  doi     = {10.5555/3291125.3309617}
+  url     = {http://jmlr.org/papers/v19/18-158.html}
 }
 
 
@@ -182,7 +182,8 @@ author = {Dmitriy Drusvyatskiy and Henry Wolkowicz}
   journal      = {Combinatorial and Computational Geometry},
   volume       = {52},
   year         = {2005},
-  url          = {https://faculty.cc.gatech.edu/~vempala/papers/survey.pdf}
+  url          = {https://faculty.cc.gatech.edu/~vempala/papers/survey.pdf},
+  doi          = {10.1017/9781009701259.033}
 }
 
 @article{lovasz1999,
@@ -191,7 +192,8 @@ author = {Dmitriy Drusvyatskiy and Henry Wolkowicz}
   journal      = {Mathematical Programming},
   volume       = {86},
   year         = {1999},
-  url          = {https://link.springer.com/content/pdf/10.1007/s101070050099.pdf}
+  url          = {https://link.springer.com/content/pdf/10.1007/s101070050099.pdf},
+  doi          = {10.1007/s101070050099}
 }
 
 @article{Simonovits2003,
@@ -200,7 +202,8 @@ author = {Dmitriy Drusvyatskiy and Henry Wolkowicz}
   journal      = {Mathematical Programming},
   volume       = {97},
   year         = {2003},
-  url          = {https://link.springer.com/article/10.1007/s10107-003-0447-x}
+  url          = {https://link.springer.com/article/10.1007/s10107-003-0447-x},
+  doi          = {10.1007/s10107-003-0447-x}
 }
 
 
@@ -238,7 +241,7 @@ number = {2},
 pages = {495-510},
 year = {2023},
 issn = {0377-2217},
-doi = {https://doi.org/10.1016/j.ejor.2023.03.021},
+doi = {10.1016/j.ejor.2023.03.021},
 url = {https://www.sciencedirect.com/science/article/pii/S037722172300228X},
 author = {Haesol Im and Henry Wolkowicz},
 keywords = {Linear programming, Facial reduction, Preprocessing, Degeneracy, Implicit problem singularity},
@@ -252,7 +255,7 @@ volume = {109},
 pages = {101916},
 year = {2023},
 issn = {0925-7721},
-doi = {https://doi.org/10.1016/j.comgeo.2022.101916},
+doi = {10.1016/j.comgeo.2022.101916},
 url = {https://www.sciencedirect.com/science/article/pii/S0925772122000591},
 author = {Ludovic Calès and Apostolos Chalkis and Ioannis Z. Emiris and Vissarion Fisikopoulos},
 keywords = {Polytope volume, Convex body, Simplex, Sampling, Empirical finance},
@@ -297,5 +300,6 @@ abstract = {We examine volume computation of general-dimensional polytopes and m
   pages = {100659},
   title = {{Polyrun: A Java library for sampling from the bounded convex polytopes}},
   volume = {13},
-  year = {2021}
+  year = {2021},
+  doi = {10.1016/j.softx.2021.100659}
 }
