+++ paper/paper.bib
@@ -121,7 +121,7 @@
 
 @inproceedings{10.5555/3600270.3602567,
 author = {Kook, Yunbum and Lee, Yin Tat and Shen, Ruoqi and Vempala, Santosh S.},
-title = {Sampling with riemannian Hamiltonian Monte Carlo in a constrained space},
+title = {{Sampling with riemannian Hamiltonian Monte Carlo in a constrained space}},
 year = {2022},
 isbn = {9781713871088},
 publisher = {Curran Associates Inc.},
